% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spp_frc_data.R
\name{calc_spp_frc_data}
\alias{calc_spp_frc_data}
\title{Calculate fractional coverage data}
\usage{
calc_spp_frc_data(
  x,
  res,
  output_dir,
  template_data = NULL,
  cache_dir = tempdir(),
  force = FALSE,
  n_threads = 1,
  engine = "terra",
  cache_limit = 1000,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::st_sf()}} Object containing the species data.
This object should be produced using the \code{\link[=create_spp_aoh_data]{create_spp_aoh_data()}}
function.}

\item{res}{\code{numeric} Resolution for computing fractional coverage.
Note that the argument to \code{res} must be a factor of the
the resolution of the underlying Area of Habitat data.
For example, a value of 5000 would be a valid argument
if the underlying data had a resolution of 100 m.}

\item{output_dir}{\code{character} \code{character} Folder path to save raster
(GeoTIFF) files containing the fractional coverage data.}

\item{template_data}{\code{\link[terra:rast]{terra::rast()}} Raster data to use as a template
for computing fractional coverage.
Note that the argument should have the same spatial properties
as the elevation and habitat data used to generate the
Area of Habitat data.
Defaults to \code{NULL} such that template data are automatically imported
as the default global habitat dataset (using
\code{\link[=get_lumbierres_habitat_data]{get_lumbierres_habitat_data()}}).}

\item{cache_dir}{\code{character} Folder path for downloading and caching data.
By default, a temporary directory is used (i.e., \code{tempdir()}).
\strong{To avoid downloading the same data multiple times, it is strongly
recommended to specify a persistent storage location (see Examples below).}}

\item{force}{\code{logical} Should the data be downloaded even if the
the data are already available?
Defaults to \code{FALSE}.}

\item{n_threads}{\code{integer} Number of computational threads to use
for data processing.
To reduce run time, it is strongly recommended to set this
parameter based on available resources (see Examples section below).
Note that parallel processing is only used for processing the
habitat classification and elevation data.
As such, this parameter will have no influence when using preprocessed
datasets.
Defaults to 1.}

\item{engine}{\code{character} Value indicating the name of the software
to use for data processing.
Available options include \code{"terra"} or \code{"gdal"}.
Defaults to \code{"terra"}.}

\item{cache_limit}{\code{integer} Amount of memory (Mb) for caching
when processing spatial data with the
Geospatial Data Abstraction Library (GDAL).
This parameter is only used when using the \code{"gdal"} engine.
If possible, it is recommended to set this as parameter to
at 5000 (assuming there is at least 8Gb memory available
on the system).
Defaults to 1000.}

\item{verbose}{\code{logical} Should progress be displayed while processing data?
Defaults to \code{TRUE}.}

\item{...}{Arguments passed to \code{\link[=get_lumbierres_habitat_data]{get_lumbierres_habitat_data()}}.}
}
\value{
An updated version of the argument to \code{x} with updated values
for the \code{path}, \code{xmin}, \code{xmax}, \code{ymin}, and \code{ymax} columns.
The
}
\description{
Calculate fractional coverage of species' Area of Habitat.
}
\details{
This function works by
(i) creating a template grid based on the specified resolution,
(ii) aggregating the Area of Habitat data to the match the
spatial origin and resolution of the template grid, and
(iii) dividing the aggregated values based on the aggregation
factor to express values as fractional coverage.
}
\section{Output file format}{

Fractional coverage data are stored in a separate raster (GeoTIFF) file for
each seasonal distribution of each species. Each raster file is assigned a
file name based on a prefix and a combination of the species' taxon
identifier
(per \code{id_no}/\code{SISID} in \code{x}) and the identifier for the seasonal
distribution (per \code{seasonality} in \code{x})
(i.e., file names are named according to \verb{FRC_\{$id_no\}_$\{seasonality\}.tif}).
For a given raster file, grid cell values denote the proportion of suitable
habitat located within each cell.
For example, a value of 0 corresponds to 0\% fractional coverage,
0.5 to 50\% fractional coverage, 1 to 100\% fractional coverage.
Missing (\code{NA}) values correspond to
grid cells that are located entirely outside of the species' distribution.
}

\examples{
\dontrun{
# find file path for example range data following IUCN Red List data format
## N.B. the range data were not obtained from the IUCN Red List,
## and were instead based on data from GBIF (https://www.gbif.org/)
path <- system.file("extdata", "EXAMPLE_SPECIES.zip", package = "aoh")

# import data
spp_range_data <- read_spp_range_data(path)

# specify settings for data processing
output_aoh_dir <- tempdir()                   # folder to save AOH data
output_frc_dir <- tempdir()                   # folder to save coverage data
cache_dir <- rappdirs::user_data_dir("aoh")   # persistent storage location
n_threads <- parallel::detectCores() - 1      # speed up analysis

# create cache directory if needed
if (!file.exists(cache_dir)) {
  dir.create(cache_dir, showWarnings = FALSE, recursive = TRUE)
}

# create species information data
spp_info_data <- create_spp_info_data(
  x = spp_range_data,
  cache_dir = cache_dir
)

# create Area of Habitat data for species
spp_aoh_data <- create_spp_aoh_data(
  x = spp_info_data,
  output_dir = output_aoh_dir,
  n_threads = n_threads,
  cache_dir = cache_dir
)

# compute fractional coverage across a 5 x 5 km spatial grid
spp_aoh_frc_data <- calc_spp_frc_data(
  x = spp_aoh_data,
  output_dir = output_frc_dir,
  res = 5000,
  cache_dir = cache_dir
)
}

\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
# preview data
print(spp_aoh_frc_data)
}
\dontshow{\}) # examplesIf}
\dontrun{
# plot the data to visualize the range maps and fractional coverage data
plot_spp_frc_data(spp_aoh_frc_data)
}
}
\seealso{
This function is useful for creating fractional coverage data when
you have previously generated species' Area of Habitat data.
If you have not previously generated species' Area of Habitat data,
you can use the \code{\link[=create_spp_frc_data]{create_spp_frc_data()}} to create fractional coverage
data directly.
}
