% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spp_aoh_data.R
\name{create_spp_aoh_data}
\alias{create_spp_aoh_data}
\title{Create Area of Habitat data}
\usage{
create_spp_aoh_data(
  x,
  output_dir,
  spp_summary_data = NULL,
  spp_habitat_data = NULL,
  elevation_data = NULL,
  habitat_data = NULL,
  template_data = get_world_berhman_1km_rast(),
  cache_dir = tempdir(),
  iucn_version = "latest",
  habitat_version = "latest",
  key = NULL,
  force = FALSE,
  parallel_n_threads = 1,
  parallel_cluster = NULL,
  omit_habitat_codes = default_omit_iucn_habitat_codes(),
  use_ee = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::sf()}} Spatial data delineating species geographic ranges
obtained from the \href{https://www.iucnredlist.org/}{IUCN Red List}.}

\item{output_dir}{\code{character} folder path to save raster files
(GeoTIFF format) containing the Area of Habitat data.}

\item{spp_summary_data}{\code{\link[tibble:tibble]{tibble::tibble()}} Table containing summary
information for each species (in the argument to \code{x}).
Specifically, the argument should contain the following columns: \code{"id_no"},
\code{"elevation_lower"}, and \code{"elevation_upper"} columns.
Here, \code{"id_no"} corresponds to the species' taxon identifier
(also present in \code{x}), and the \code{"elevation_lower"} and \code{"elevation_upper"}
columns indicate the lowest and highest elevations that contain habitat
for the species.
Defaults to \code{NULL} such that data are automatically obtained from the
latest version of the \href{https://www.iucnredlist.org}{IUCN Red List}.}

\item{spp_habitat_data}{\code{\link[tibble:tibble]{tibble::tibble()}} Table containing habitat
preference information for each species (in the argument to \code{x}).
Specifically, the argument should contain the following columns: \code{"id_no"},
\code{"code"}, \code{"suitability"}, \code{"season"} columns.
Here, \code{"id_no"} corresponds to the species' taxon identifier
(also present in \code{x}), \code{"code"} indicates a habitat classification code
that is suitable for the species (i.e. based on layer names in the
argument to habitat_data), \code{"suitability"} indicates the level suitability
of the habitat class for a given species (e.g. using values such
as \code{"Suitable"} or \code{"Marginal"}), and \code{"season"} indicates
if the habitat class is only suitable for a particular seasonal
distribution (e.g. \code{"Breeding"}).
Defaults to \code{NULL} such that data are automatically obtained from the
latest version of the \href{https://www.iucnredlist.org}{IUCN Red List}.}

\item{elevation_data}{\code{\link[terra:rast]{terra::rast()}} Raster data delineating the
worldwide elevation data (e.g. Jung \emph{et al.} 2020a).
Defaults to \code{NULL} such that data
are automatically obtained from Jung \emph{et al.} 2020b
(using \code{\link[=get_global_elevation_data]{get_global_elevation_data()}}).}

\item{habitat_data}{\code{\link[terra:rast]{terra::rast()}} Multi-layer raster data delineating the
coverage of different habitat classes across world
(e.g. Jung \emph{et al.} 2020a).
Each layer should correspond to a different habitat class,
and the name of each layer should contain a unique code used to identify
which places contain suitable habitat for a given species
(per the \code{"code"} column in the the argument to \code{spp_habitat_data}).
For each layer, pixel values should indicate the percentage of available
an available habitat class.
\strong{Critically, a value of 1000 should indicate 100\%, and a value of 0
should indicate 0\%.}
Defaults to \code{NULL} such that data
are automatically obtained from Jung \emph{et al.} 2020b
(using \code{\link[=get_global_habitat_data]{get_global_habitat_data()}}).}

\item{template_data}{\code{\link[terra:rast]{terra::rast()}} Template raster specifying the
resolution, coordinate reference system, and dimensionality for the
output raster files.
Defaults to \code{\link[=get_world_berhman_1km_rast]{get_world_berhman_1km_rast()}} which is a global 1 km resolution
raster under the World Behrman coordinate reference system (ESRI:54017).}

\item{cache_dir}{\code{character} folder path for downloading and caching data.
By default, a temporary directory is used (i.e. \code{tempdir()}).
\strong{To avoid downloading the same data multiple times, it is strongly
recommended to specify a persistent storage location (see Examples below).}}

\item{iucn_version}{\code{character} Version of the
IUCN Red List dataset that should be used. See documentation for the
the \code{version} parameter in the \code{\link[=get_spp_summary_data]{get_spp_summary_data()}} function
for further details.
Defaults to \code{"latest"} such that the most recent version of the dataset is
used.}

\item{habitat_version}{\code{character} Version of the
habitat dataset that should be used. See documentation for the
the \code{version} parameter in the \code{\link[=get_global_habitat_data]{get_global_habitat_data()}} function
for further details.
Defaults to \code{"latest"} such that the most recent version of the dataset is
used.}

\item{key}{\code{character} Token for querying the IUCN Red List API.
\strong{Please note that a token is required to download
data from the IUCN Red List.} To obtain a token, please see
\code{\link[rredlist:rl_use_iucn]{rredlist::rl_use_iucn()}}. Defaults to \code{NULL} such that the token
accessed from  the \code{"IUCN_REDLIST_KEY"} environmental variable
(which can be specified in the .Renviron file).}

\item{force}{\code{logical} should the data be downloaded even if the
the data are already available?
Defaults to \code{FALSE}.}

\item{parallel_n_threads}{\code{integer} Number of computational threads to use
for data processing.
\strong{To reduce run time, it is strongly recommended to set this
parameter based on the number of available threads (see Examples below).}
Note that this parameter is only used if \code{use_ee} is \code{FALSE}.
Defaults to 1.}

\item{parallel_cluster}{\code{character} Name of strategy for
processing data in parallel. Available options are \code{"FORK"} and
\code{"PSOCK"}.
Defaults to \code{NULL} such that \code{"PSOCK"} is used on Microsoft
Windows operating systems, and \code{"FORK"} otherwise.
Note that this parameter only affects data pre-processing if
Earth Engine is used (i.e. if \code{use_ee} is \code{TRUE}).
Defaults to 1.}

\item{omit_habitat_codes}{\code{character} Habitat classification codes
to omit from resulting Area of Habitat data. If the aim is to identify
places that contain natural, suitable habitats, then processing should
exclude (i) anthropogenically modified habitat classifications and (ii)
unknown habitat classification types. Defaults to all artificial,
introduced vegetation, and unknown habitat classifications on
the \href{https://www.iucnredlist.org/resources/habitat-classification-scheme}{IUCN Red List Habitat Classification Scheme}
(see \code{\link[=default_omit_iucn_habitat_codes]{default_omit_iucn_habitat_codes()}},}

\item{use_ee}{\code{logical} Should data processing be completed using
Google Earth Engine (via the \pkg{rgee}) package?
Defaults to \code{FALSE}.}

\item{verbose}{\code{logical} Should progress be displayed while processing data?
Defaults to \code{TRUE}.}
}
\value{
A \code{\link[sf:sf]{sf::st_sf()}} object containing range maps for the species distributions
used to generate the Area of Habitat data and additional columns describing
the Area of Habitat data.
These range maps were produced by cleaning those supplied as an argument
to \code{x} so they can be used to generate the Area of Habitat data
(see \code{\link[=clean_spp_range_data]{clean_spp_range_data()}} for data cleaning procedures).
Specifically, the object contains the following columns:
\describe{
\item{id_no}{\code{numeric} species' taxon identifier on the IUCN Red List.}
\item{seasonal}{\code{numeric} seasonal distribution code.}
\item{habitat_code}{\code{character} habitat
codes used to create the species' Area of Habitat data.
If a given species has multiple suitable habitat classes,
then these are denoted using a pipe-delimeted format.
For example, if the habitat classes denoted with the codes
\code{"1.5"} and \code{"1.9"} were considered suitable for a given species, then
these codes would be indicated as \code{"1.5|1.9"}.}
\item{elevation_lower}{\code{numeric} lower elevation threshold used to create
the species' Area of Habitat data.}
\item{elevation_upper}{\code{numeric} upper elevation threshold used to create
the species' Area of Habitat data.}
\item{xmin}{\code{numeric} value describing the spatial extent of
the Area of Habitat data.}
\item{xmax}{\code{numeric} value describing the spatial extent of
the Area of Habitat data.}
\item{ymin}{\code{numeric} value describing the spatial extent of
the Area of Habitat data.}
\item{ymax}{\code{numeric} value describing the spatial extent of
the Area of Habitat data.}
\item{path}{\code{character} file paths for Area of Habitat data
(in GeoTIFF format).
The Area of Habitat datasets denote the proportion of land
inside different grid cells that contain suitable habitat for
species' seasonal distribution.
Thus a grid cell with a value of 0 has 0\% of its spatial extent covered by
suitable habitat, a value of 1 has 100\% of its spatial extent covered by
means that the cell is located outside of the species seasonal
distribution (per the geographic range data).
File paths that are denoted with missing (\code{NA}) values correspond to
species distributions that were not processed (e.g. due to lack of habitat
data).
}
}
}
\description{
Create Area of Habitat (AOH) data for species based on their altitudinal and
habitat preferences (Brooks \emph{et al.} 2019). This function is designed to
work with data obtained from the
\href{https://www.iucnredlist.org/}{International Union for Conservation of Nature (IUCN) Red List of Threatened Species}.
}
\examples{
\dontrun{
# load built in datasets
data(sim_spp_summary_data, package = "aoh")
data(sim_spp_habitat_data, package = "aoh")

# find file path for simulated data following the IUCN Red List format
path <- system.file("extdata", "SIMULATED_SPECIES.zip", package = "aoh")

# import species range data
sim_spp_range_data <- read_spp_range_data(path)

# specify settings for data processing
output_dir <- tempdir()                  # folder to save AOH data
cache_dir <- rappdirs::user_data_dir()   # persistent storage location
n_threads <- parallel::detectCores() - 1 # recommended for best performance

# create Area of Habitat data for species
sim_aoh_data <- create_aoh_data(
  x = sim_spp_range_data,
  output_dir = output_dir,
  spp_summary_data = sim_spp_summary_data,
  spp_habitat_data = sim_spp_habitat_data,
  parallel_n_threads = n_threads,
  cache_dir = cache_dir
)

# preview result
print(sim_aoh_data)

# import AOH data
aoh_data <- lapply(sim_aoh_data, terra::rast)

# compare range data (EOO) and AOH data for first species
#TODO
}
}
\references{
Amatulli G, Domisch S, Tuanmu M-N, Parmentier B, Ranipeta A, Malczyk J, and
Jetz W (2018) A suite of global, cross-scale topographic variables for
environmental and biodiversity modeling. Scientific Data, 5:180040.
\url{https://doi.org/10.1038/sdata.2018.40}

Brooks TM, Pimm SL, Akçakaya HR, Buchanan GM, Butchart SHM, Foden W,
Hilton-Taylor C, Hoffmann M, Jenkins CN, Joppa L, Li BV, Menon V,
Ocampo-Peñuela N, Rondinini C (2019) Measuring terrestrial Area of Habitat
(AOH) and its Utility for the IUCN Red List. Trends in Ecology & Evolution.
34:977--986. \url{doi:10.1016/j.tree.2019.06.009}

Jung M, Dahal PR, Butchart SH, Donald PF, De Lamo X, Lesiv M, Kapos V,
Rondinini C, and Visconti P (2020a) A global map of
terrestrial habitat types. Scientific data, 7:1--8.
\url{https://doi.org/10.1038/s41597-020-00599-8}

Jung M, Dahal PR, Butchart SH, Donald PF, De Lamo X, Lesiv M, Kapos V,
Rondinini C, and Visconti P (2020b) A global map of
terrestrial habitat types (insert version) [Data set]. Zenodo.
\url{https://doi.org/10.5281/zenodo.4058819}
}
