% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spp_info_data.R
\name{create_spp_info_data}
\alias{create_spp_info_data}
\title{Create species' information data}
\usage{
create_spp_info_data(
  x,
  spp_summary_data = NULL,
  spp_habitat_data = NULL,
  cache_dir = tempdir(),
  iucn_version = "latest",
  key = NULL,
  force = FALSE,
  keep_iucn_rl_presence = c(1, 2),
  keep_iucn_rl_origin = c(1, 2, 6),
  keep_iucn_rl_seasonal = c(1, 2, 3, 4),
  omit_habitat_codes = iucn_habitat_codes_marine(),
  adjust_elevational_limits = TRUE,
  adjust_habitat_codes = TRUE,
  crs = sf::st_crs("ESRI:54017"),
  geometry_precision = 1e+05,
  verbose = TRUE
)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::sf()}} Spatial data delineating species geographic ranges
obtained from the \href{https://www.iucnredlist.org/}{IUCN Red List}.
See below for details.}

\item{spp_summary_data}{\code{\link[tibble:tibble]{tibble::tibble()}} Table containing summary
information for each species (in the argument to \code{x}).
Specifically, the argument should contain the following columns: \code{"id_no"},
\code{"elevation_lower"}, and \code{"elevation_upper"} columns.
Here, \code{"id_no"} corresponds to the species' taxon identifier
(also present in \code{x}), and the \code{"elevation_lower"} and \code{"elevation_upper"}
columns indicate the lowest and highest elevations that contain habitat
for the species.
Defaults to \code{NULL} such that data are automatically obtained from the
latest version of the \href{https://www.iucnredlist.org}{IUCN Red List}.}

\item{spp_habitat_data}{\code{\link[tibble:tibble]{tibble::tibble()}} Table containing habitat
preference information for each species (in the argument to \code{x}).
Specifically, the argument should contain the following columns: \code{"id_no"},
\code{"code"}, \code{"suitability"}, \code{"season"} columns.
Here, \code{"id_no"} corresponds to the species' taxon identifier
(also present in \code{x}), \code{"code"} indicates a habitat classification code
that is suitable for the species (i.e., based on layer names in the
argument to habitat_data), \code{"suitability"} indicates the level suitability
of the habitat class for a given species (e.g., using values such
as \code{"Suitable"} or \code{"Marginal"}), and \code{"season"} indicates
if the habitat class is only suitable for a particular seasonal
distribution (e.g., \code{"Breeding"}).
Defaults to \code{NULL} such that data are automatically obtained from the
latest version of the \href{https://www.iucnredlist.org}{IUCN Red List}.}

\item{cache_dir}{\code{character} Folder path for downloading and caching data.
By default, a temporary directory is used (i.e., \code{tempdir()}).
\strong{To avoid downloading the same data multiple times, it is strongly
recommended to specify a persistent storage location (see Examples below).}}

\item{iucn_version}{\code{character} Version of the
IUCN Red List dataset that should be used. See documentation for the
the \code{version} parameter in the \code{\link[=get_spp_summary_data]{get_spp_summary_data()}} function
for further details.
Defaults to \code{"latest"} such that the most recent version of the dataset is
used.}

\item{key}{\code{character} Token for querying the IUCN Red List API.
Defaults to \code{NULL} such that the token
accessed from  the \code{"IUCN_REDLIST_KEY"} environmental variable
(which can be specified in the .Renviron file).}

\item{force}{\code{logical} Should the data be downloaded even if the
the data are already available?
Defaults to \code{FALSE}.}

\item{keep_iucn_rl_presence}{\code{integer} IUCN Red List presence codes
to retain
(see IUCN SSC Red List Technical Working Group 2021 for details).
Species' ranges that are not associated with these codes are
excluded during data cleaning procedures.
Defaults to a numeric vector containing \code{1} and \code{2}
(corresponding to \emph{extant} and \emph{probably extant}).}

\item{keep_iucn_rl_origin}{\code{integer} IUCN Red List origin codes
to retain
(see IUCN SSC Red List Technical Working Group 2021 for details).
Species' ranges that are not associated with these codes are
excluded during data cleaning procedures.
Defaults to a numeric vector containing \code{1}, \code{2}, and \code{6}.
(corresponding to \emph{native}, \emph{reintroduced}, and \emph{assisted colonization}).}

\item{keep_iucn_rl_seasonal}{\code{integer} IUCN Red List seasonal codes
to retain
(see IUCN SSC Red List Technical Working Group 2021 for details).
Species' ranges that are not associated with these codes are
excluded during data cleaning procedures.
Defaults to a numeric vector containing \code{1}, \code{2}, \code{3}, and \code{4}.
(corresponding to \emph{resident}, \emph{breeding season}, \emph{non-breeding season},
and \emph{passage} distributions).}

\item{omit_habitat_codes}{\code{character} Habitat classification codes
to omit from resulting Area of Habitat data.
Please see the \href{https://www.iucnredlist.org/resources/habitat-classification-scheme}{IUCN Red List Habitat Classification Scheme}
for the full range of habitat classification codes.
For example,
if the aim is to identify natural places that contain suitable conditions,
then areas classified as anthropogenically modified
(\code{\link[=iucn_habitat_codes_artificial]{iucn_habitat_codes_artificial()}}),
introduced vegetation (\code{\link[=iucn_habitat_codes_introduced]{iucn_habitat_codes_introduced()}},
or unknown habitat (\code{\link[=iucn_habitat_codes_misc]{iucn_habitat_codes_misc()}}) should
be excluded.
Defaults to \code{\link[=iucn_habitat_codes_marine]{iucn_habitat_codes_marine()}}, such that marine
habitats are excluded.}

\item{adjust_elevational_limits}{\code{logical} Should elevation limits be
adjusted to correct for errors? Defaults to \code{TRUE} to automatically
fix errors (see Data processing section below for details).}

\item{adjust_habitat_codes}{\code{logical} Should habitat codes be adjusted so
they are assigned to species' distributions following guidelines for Key
Biodiversity areas?
Defaults to \code{TRUE} (see Data processing section below for details).
Otherwise, if \code{FALSE}, habitat codes are assigned to  species'
distributions following
exact matches (e.g., only codes described for a \emph{resident} distribution
are assigned to the \emph{resident} distribution).}

\item{crs}{A \code{st_crs()} object containing the coordinate reference system
for reprojecting the species' range data.
This coordinate reference system should be the same as the elevation
and habitat classification data that will subsequently be used to
generate Area of Habitat data (using the \code{\link[=create_spp_aoh_data]{create_spp_aoh_data()}} or
\code{\link[=create_spp_frc_data]{create_spp_frc_data()}} functions).
Defaults to the World Behrmann coordinate reference system (ESRI:54017),
which is appropriate for the elevation and habitat datasets
available through the package
(e.g., using the \code{\link[=get_global_elevation_data]{get_global_elevation_data()}} and
\code{\link[=get_lumb_cgls_habitat_data]{get_lumb_cgls_habitat_data()}} functions).}

\item{geometry_precision}{\code{numeric} level of precision for processing the
spatial data (used with \code{\link[sf:st_precision]{sf::st_set_precision()}}). The default is
100000 (higher values indicate higher precision).
Although this level of precision is generally suitable for fine-scale
analyses, it might result in unnecessarily long computation times
(e.g., 1500 is suitable for national-scale analyses). If you encounter
geometry errors, increasing the argument to the
parameter can sometimes resolve these issues.}

\item{verbose}{\code{logical} Should progress be displayed while processing data?
Defaults to \code{TRUE}.}
}
\value{
A \code{\link[sf:sf]{sf::st_sf()}} object containing all the information needed
to create Area of Habitat data. It contains
cleaned versions of the species' range maps and columns containing
the species' elevational limits and habitat preferences.
Specifically, the object contains the following columns:
\describe{
\item{id_no}{\code{numeric} species' taxon identifier on the IUCN Red List.}
\item{binomial}{\code{character} species name.}
\item{category}{\code{character} IUCN Red List threat category.}
\item{migratory}{\code{logical} indicating if the species was processed
as a migratory species (i.e., it had a \emph{breeding}, \emph{non-breeding}, or
\emph{passage} seasonal distribution).}
\item{seasonal}{\code{numeric} seasonal distribution code.}
\item{full_habitat_code}{\code{character} all habitat classification
codes that contain suitable habitat for the species.
If a given species has multiple suitable habitat classes,
then these are denoted using a pipe-delimited format.
For example, if the habitat classes denoted with the codes
\code{"1.5"} and \code{"1.9"} were considered suitable for a given species, then
these codes would be indicated as \code{"1.5|1.9"}.}
\item{elevation_lower}{\code{numeric} lower elevation threshold used to create
the species' Area of Habitat data.}
\item{elevation_upper}{\code{numeric} upper elevation threshold used to create
the species' Area of Habitat data.}
\item{geometry}{\code{sf::st_sfc()} geometries for species' distributions.}
}
}
\description{
Create data to collate all the information needed to generate
Area of Habitat data (Brooks \emph{et al.} 2019).
Briefly, this function cleans and collates data describing species'
geographic ranges, habitat preferences, and elevational limits.
Please note that these procedures are designed for terrestrial species,
and will not apply to marine or freshwater species.
}
\section{Species range data format}{

Species range data are expected to follow the data format conventions
for the IUCN Red List (see \href{https://www.iucnredlist.org/resources/mappingstandards}{IUCN Red List documentation} for
details). Specifically, the argument to \code{x} should be an
\code{\link[sf:sf]{sf::st_sf()}} object with the following columns: \code{id_no},
\code{presence}, \code{origin}, and \code{seasonal}.
It can also contain the following optional columns: \code{terrestrial}
(or \code{terrestial}), \code{freshwater}, and \code{marine}.
Below we provide a brief description of each column:

\describe{

\item{\code{id_no}}{\code{numeric} taxon identifier on the IUCN Red List.}

\item{\code{presence}}{\code{numeric} identifier describing information about
the presence of the taxon in the range data.}

\item{\code{origin}}{\code{numeric}  identifier describing if the species is native
to the location(s) described by the range data.}

\item{\code{seasonality}}{\code{numeric} identifier describing if the species
is occupied by the location(s) describe by the range data throughout
the whole year, of if only during certain seasons.}

\item{\code{terrestial}}{\code{character} value indicating if the range
data pertain to terrestrial environments (with \code{"true"} or \code{"false"}
values.)}

\item{\code{freshwater}}{\code{character} value indicating if the range
data pertain to freshwater environments (with \code{"true"} or \code{"false"}
values.)}

\item{\code{marine}}{\code{character} value indicating if the range
data pertain to marine environments (with \code{"true"} or \code{"false"}
values.)}

}
}

\section{Data processing}{

The species' information data are produced using the following procedures.
\enumerate{
\item Species range data cleaned. By default, the range data are cleaned
following guidelines for the identification of Key Biodiversity Areas
(KBA Standards and Appeals Committee of IUCN SSC/WCPA 2022).
Specifically, the default cleaning procedures involves excluding places
where the
(i) species' presence is not \emph{extant} or
\emph{probably extant}
(i.e. filtering based on \code{presence == 1} or \code{presence == 2});
(ii) species' origin is not \emph{native},
\emph{reintroduced}, or the result of \emph{assisted colonization}
(i.e. filtering based on \code{origin == 1}, \code{origin == 2}, or \code{origin == 6});
(iii) available information on which species' seasonal distribution
is \emph{uncertain} (i.e. filtering based on \code{seasonal != 5}); and
(iv) species' distribution is not terrestrial
(i.e. filtering based on where \code{terrestrial == "true"}).
Additionally, the species' range data are spatially dissolved so that each
seasonal distribution for each taxon is represented by a separate geometry.
Finally, geoprocessing routines are used to detect and repair
any invalid geometries.
\item Species summary and habitat preferences data are
imported (if needed, see \code{\link[=get_spp_summary_data]{get_spp_summary_data()}} and
\code{\link[=get_spp_habitat_data]{get_spp_habitat_data()}} for details).
If these data are not available in the cache directory
(i.e. argument to \code{cache_dir}), then they are automatically downloaded
to the cache directory.
\item If specified (per \code{adjust_elevational_limits = TRUE}),
then the elevational limit values in the species summary data
are adjusted to correct for errors. These adjustments applied based on
the following procedures:
(i) if a species lacks lower or upper elevational limits,
then limits of -500 m and 9,000 m are assumed (respectively);
(ii) lower elevational limit values below -500 m are replaced with -500 m;
(iii) upper elevational limit values above 9000 are is replaced with
9000 m;
(iv) if a lower elevational limit is greater than an upper elevational
limit, then limits of -500 m and 9000 m are assumed (respectively);
(v) if a lower elevational limit is within 50 m of an
upper elevational limit, then limits are adjusted such that there is a
50 m difference between them.
Otherwise, if \code{FALSE}, then elevation limit values are not altered.
\item Each is species is classified as either migratory or non-migratory,
based on the presence of \emph{breeding}, \emph{non-breeding}, or \emph{passage}
distributions in the species range data (i.e., x`). For example,
if a species only has a \emph{resident} distribution in the species range
data, then it is classified as a non-migratory species. If a
species has a \emph{resident} and a \emph{breeding} distribution in the
species range data, then it is classified as a migratory species.
\item If specified (per \code{adjust_habitat_codes = TRUE}),
then the habitat codes in the species habitat preferences data
are adjusted based on guidelines for the identification of
Key Biodiversity Areas
(KBA Standards and Appeals Committee of IUCN SSC/WCPA 2022).
These adjustments are based on the following procedures:
(i) \emph{resident} distributions for non-migratory species are assigned habitat
codes described in the species habitat preference data for the species'
\emph{resident}, \emph{breeding}, \emph{non-breeding}, \emph{passage},
\emph{seasonal occurrence uncertain}, and missing (\code{NA}) seasonal distributions;
(ii) \emph{resident} distributions for migratory species are assigned habitat
codes described in the species habitat preference data for the species'
\emph{resident}, \emph{breeding}, \emph{non-breeding}, \emph{seasonal occurrence uncertain}, and
missing (\code{NA}) seasonal distributions;
(ii) \emph{breeding} distributions are assigned habitat codes
described for the species' \emph{resident}, \emph{breeding},
\emph{seasonal occurrence uncertain} and missing (\code{NA}) seasonal distributions;
(iii) \emph{non-breeding} distributions are assigned
habitat codes described for the species' \emph{resident}, \emph{non-breeding},
\emph{seasonal occurrence uncertain}, and missing (\code{NA})
seasonal distributions; and
(iv) \emph{passage} distributions are assigned habitat codes described for the
species' \emph{resident}, \emph{passage}, \emph{seasonal occurrence uncertain}, and
missing (\code{NA}) seasonal distributions.
If the adjustments are not applied (per \code{adjust_habitat_codes = FALSE}),
then the habitat codes are assigned to species' distributions based on
exact matches (e.g., only codes described for a \emph{resident} distribution
are assigned to the \emph{resident} distribution).
\item Species information are collated into a single dataset containing
their geographic ranges, migratory status, habitat preferences, and
elevational limits.
Specifically, taxon identifiers (per the \code{id_no}/\code{SISID} columns)
are used merge the datasets together.
\item Post-processing routines are used to prepare the output data.
}
}

\examples{
\dontrun{
# find file path for example range data following IUCN Red List data format
## N.B., the range data were not obtained from the IUCN Red List,
## and were instead based on data from GBIF (https://www.gbif.org/)
path <- system.file("extdata", "EXAMPLE_SPECIES.zip", package = "aoh")

# import data
spp_range_data <- read_spp_range_data(path)

# specify persistent storage location for data processing
cache_dir <- rappdirs::user_data_dir("aoh")

# create cache directory if needed
if (!file.exists(cache_dir)) {
  dir.create(cache_dir, showWarnings = FALSE, recursive = TRUE)
}

# create species information data
spp_info_data <- create_spp_info_data(
  x = spp_range_data,
  cache_dir = cache_dir
)
}

\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
# preview data
print(spp_info_data)
}
\dontshow{\}) # examplesIf}
}
\references{
Brooks TM, Pimm SL, Akçakaya HR, Buchanan GM, Butchart SHM, Foden W,
Hilton-Taylor C, Hoffmann M, Jenkins CN, Joppa L, Li BV, Menon V,
Ocampo-Peñuela N, Rondinini C (2019) Measuring terrestrial Area of Habitat
(AOH) and its utility for the IUCN Red List. \emph{Trends in Ecology & Evolution},
34, 977--986. \doi{10.1016/j.tree.2019.06.009}

KBA Standards and Appeals Committee of IUCN SSC/WCPA
(2022). \emph{Guidelines for using A Global Standard for the Identification}
\emph{of Key Biodiversity Areas}. Version 1.2. Gland, Switzerland: IUCN.
}
