% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_terra.R
\name{create_template_rast}
\alias{create_template_rast}
\title{Create a template raster}
\usage{
create_template_rast(xres, yres, crs, bbox)
}
\arguments{
\item{xres}{\code{numeric} Resolution for the x-axis.}

\item{yres}{\code{numeric} Resolution for the y-axis.}

\item{crs}{\code{\link[sf:st_crs]{sf::st_crs()}} Spatial reference coordinate system object.}

\item{bbox}{\code{\link[sf:st_bbox]{sf::st_bbox()}} Spatial bounding box object.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} object with desired spatial properties.
}
\description{
Create a template \code{\link[terra:rast]{terra::rast()}} raster object based on specified
spatial properties.
}
\details{
Please not that all grid cells contain missing (\code{NA}) values.
}
\examples{
# load data
nc <- read_sf(system.file("shape/nc.shp", package = "sf"))

# re-project to projected coordinate system
nc <- st_transform(nc, st_crs("ESRI:54017"))

# create blank raster from nc
r <-  create_template_rast(1000, 1000, st_crs(nc), st_bbox(nc))

# preview raster
print(r)

}
\seealso{
Other geoprocessing: 
\code{\link{intersecting_ext}()},
\code{\link{read_sf_n}()},
\code{\link{sf_terra_crs}()},
\code{\link{sf_terra_ext}()},
\code{\link{terra_st_bbox}()},
\code{\link{terra_st_crs}()}
}
\concept{geoprocessing}
