% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_jung_data.R
\docType{data}
\name{crosswalk_jung}
\alias{crosswalk_jung}
\alias{crosswalk_jung_data}
\title{Crosswalk data for Jung \emph{et al} 2020}
\format{
A data frame with 126 rows and 2 columns.
}
\usage{
data(crosswalk_jung)
}
\description{
TODO
}
\keyword{datasets}
