% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_jung_lvl1_data.R
\docType{data}
\name{crosswalk_jung_lvl1_data}
\alias{crosswalk_jung_lvl1_data}
\title{Crosswalk data for Jung \emph{et al.} (2020) (level 1 classification)}
\format{
A data frame (\code{\link[tibble:tibble]{tibble::tibble()}}) object with
126 rows and
2 columns.
Each row corresponds to a different IUCN habitat class.
It has the following columns:
\describe{
\item{code}{The \code{character} code for a given IUCN habitat class.}
\item{value}{The \code{numeric} value assigned to grid cells in the
raster data that contain the IUCN habitat class
(see \code{\link[=get_jung_lvl1_habitat_data]{get_jung_lvl1_habitat_data()}}).}
}
}
\source{
The data were obtained from \url{https://doi.org/10.5281/zenodo.3666245}
(Jung \emph{et al.} 2020b).
}
\usage{
data(crosswalk_jung_lvl1_data)
}
\description{
The \href{https://www.iucnredlist.org/}{International Union for Conservation of Nature (IUCN) Red List of Threatened Species}
provides a
\href{https://www.iucnredlist.org/resources/habitat-classification-scheme}{habitat classification scheme}
for different habitat types.
To map the spatial location of each of these habitat classes,
Jung \emph{et al.} (2020a) developed a global raster dataset by combining
a range of different land use and land cover data (Jung \emph{et al.} 2020b).
They created two habitat classification schemes -- termed level 1 and
level 2 classification schemes -- to describe coarse-scale and
fine-scale differences in habitat classes.
This dataset provides a crosswalk table to associate the values in the
level 1 raster dataset with IUCN habitat classes.
}
\examples{
# load data
data(crosswalk_jung_lvl1_data)

# print data
print(crosswalk_jung_lvl1_data)
}
\references{
Jung M, Dahal PR, Butchart SHM, Donald PF, De Lamo X, Lesiv M, Kapos V,
Rondinini C, and Visconti P (2020a) A global map of
terrestrial habitat types. \emph{Scientific Data}, 7:1--8.
Available at \url{https://doi.org/10.1038/s41597-020-00599-8}.

Jung M, Dahal PR, Butchart SHM, Donald PF, De Lamo X, Lesiv M, Kapos V,
Rondinini C, and Visconti P (2020b) A global map of
terrestrial habitat types (insert version) [Data set].
\emph{Zenodo Digital Repository}.
Available at \url{https://doi.org/10.5281/zenodo.4058819}.
}
\seealso{
A preprocessed version of the habitat classification data can be imported
using \code{\link[=get_jung_lvl1_habitat_data]{get_jung_lvl1_habitat_data()}}.
}
\keyword{datasets}
