% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_jung_plvl1_data.R
\docType{data}
\name{crosswalk_jung_plvl1_data}
\alias{crosswalk_jung_plvl1_data}
\title{Crosswalk data for Jung (2020) potential habitat classification (level 1)}
\format{
A data frame (\code{\link[tibble:tibble]{tibble::tibble()}}) object with
126 rows and
2 columns.
Each row corresponds to a different IUCN habitat class.
It has the following columns:
\describe{
\item{code}{The \code{character} code for a given IUCN habitat class.}
\item{value}{The \code{numeric} value assigned to grid cells in the
raster data that contain the IUCN habitat class
(see \code{\link[=get_jung_plvl1_habitat_data]{get_jung_plvl1_habitat_data()}}).}
}
}
\source{
The data were obtained from \url{https://doi.org/10.5281/zenodo.3666245}
(Jung 2020).
}
\usage{
data(crosswalk_jung_plvl1_data)
}
\description{
The \href{https://www.iucnredlist.org/}{International Union for Conservation of Nature (IUCN) Red List of Threatened Species}
provides a
\href{https://www.iucnredlist.org/resources/habitat-classification-scheme}{habitat classification scheme}
for different habitat types.
To map the spatial location of where restoration actions might produce
these habitat classes,
Jung (2020) developed a global raster dataset based on
potential vegetation data.
}
\examples{
# load data
data(crosswalk_jung_plvl1_data)

# print data
print(crosswalk_jung_plvl1_data)
}
\references{
Jung M (2020) A layer of global potential habitats (insert version)
[Data set]. \emph{Zenodo Digital Repository}.
Available at \url{https://doi.org/10.5281/zenodo.4038749}.
}
\seealso{
A preprocessed version of the habitat classification data can be imported
using \code{\link[=get_jung_plvl1_habitat_data]{get_jung_plvl1_habitat_data()}}.
}
\keyword{datasets}
