% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_lumbierres_data.R
\docType{data}
\name{crosswalk_lumbierres_data}
\alias{crosswalk_lumbierres_data}
\title{Crosswalk data for Lumbierres \emph{et al} 2021}
\format{
A data frame (\code{\link[tibble:tibble]{tibble::tibble()}}) object with
159 rows and
2 columns.
Each row corresponds to a different IUCN habitat class.
It has the following columns:
\describe{
\item{code}{The \code{character} code for a given IUCN habitat class.}
\item{value}{The \code{numeric} value assigned to grid cells in the
raster data that contain the IUCN habitat class
(see \code{\link[=get_lumbierres_habitat_data]{get_lumbierres_habitat_data()}}).}
}
}
\source{
The data were obtained from \url{https://doi.org/10.5281/zenodo.5146072}
(Lumbierres 2021).
}
\usage{
data(crosswalk_lumbierres_data)
}
\description{
The \href{https://www.iucnredlist.org/}{International Union for Conservation of Nature (IUCN) Red List of Threatened Species}
provides a
\href{https://www.iucnredlist.org/resources/habitat-classification-scheme}{habitat classification scheme}
for different habitat types.
To map the spatial location of each of these habitat classes,
Lumbierres \emph{et al.} (2021) developed a global habitat classification dataset
using the Copernicus Global Land Service Land Cover
(CGLS-LC100) dataset (Buchhorn \emph{et al.}, 2020; Buchhorn \emph{et al.}, 2019).
This dataset provides a crosswalk table to associate the values in the
raster dataset with IUCN habitat classes.
}
\examples{
# load data
data(crosswalk_lumbierres_data)

# print data
print(crosswalk_lumbierres_data)
}
\references{
Lumbierres, M (2021). Map of habitat classes (Level 1) from the IUCN
Habitat.\emph{Zenodo Digital Repository}.
Available at \url{https://doi.org/10.5281/zenodo.5146072}.

Lumbierres M, Dahal PR, Di Marco M, Butchart SHM, Donald PF, and
Rondinini C (2021) Translating habitat class to land cover to map area of
habitat of terrestrial vertebrates. \emph{Conservation Biology}, In press,
DOI:10.1111/cobi.13851
Available at \url{https://doi.org/10.1111/cobi.13851}.
}
\seealso{
A preprocessed version of the habitat classification data can be imported
using \code{\link[=get_lumbierres_habitat_data]{get_lumbierres_habitat_data()}}.
}
\keyword{datasets}
