% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_omit_iucn_habitat_codes.R
\name{default_omit_iucn_habitat_codes}
\alias{default_omit_iucn_habitat_codes}
\title{Default IUCN Red List Habitat Classification Scheme omission codes}
\usage{
default_omit_iucn_habitat_codes()
}
\value{
A \code{character} vector of habitat codes.
}
\description{
The \href{https://www.iucnredlist.org/}{International Union for Conservation of Nature (IUCN) Red List of Threatened Species}
provides a
\href{https://www.iucnredlist.org/resources/habitat-classification-scheme}{habitat classification scheme}
for different habitat types worldwide.
Under this scheme, each habitat type is assigned to a code.
This function returns the
codes for habitat types that correspond to artificial, unknown,
miscellaneous, or invasive vegetation habitat types that should be omitted
from Area of Habitat calculations.
}
\examples{
# print omission codes
print(default_omit_iucn_habitat_codes())
}
