% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_gdal.R
\name{gdal_datatype}
\alias{gdal_datatype}
\title{GDAL datatype}
\usage{
gdal_datatype(x)
}
\arguments{
\item{x}{\code{character} Value containing data type.}
}
\value{
A \code{character} containing the converted data type.
}
\description{
Convert a raster data type to the format used by GDAL.
}
