% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_available.R
\name{gdal_version}
\alias{gdal_version}
\title{GDAL version}
\usage{
gdal_version()
}
\value{
A \code{character} value describing the version of GDAL installed.
If GDAL is not installed, then a missing (\code{NA}) value is returned.
}
\description{
Find the version of the Geospatial Data Abstraction Library (GDAL)
that is currently installed.
}
\examples{
# show version of GDAL installed
print(gdal_version())

}
