% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spp_aoh_data.R
\name{get_ggmap_basemap}
\alias{get_ggmap_basemap}
\title{Get a basemap using ggmap}
\usage{
get_ggmap_basemap(x, ...)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::st_sf()}} object.}

\item{...}{Additional arguments passed to \code{\link[ggmap:get_stamenmap]{ggmap::get_stamenmap()}}.}
}
\value{
A \code{\link[ggmap:get_stamenmap]{ggmap::get_stamenmap()}} object.
}
\description{
Get a basemap for plotting spatial data using the \pkg{ggmap} package.
}
