% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_habitat_data.R
\name{get_global_habitat_data}
\alias{get_global_habitat_data}
\title{Get global habitat classification data}
\usage{
get_global_habitat_data(
  dir = tempdir(),
  version = "latest",
  preprocessed = TRUE,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dir}{\code{character} folder path for downloading and caching data.
By default, data are downloaded to a temporary directory (i.e. \code{tempdir()}).
\strong{To avoid downloading the same data multiple times, it is strongly
recommended to specify a persistent storage location (see Examples below).}}

\item{version}{\code{character} indicating the specific version of the dataset
that should be downloaded. The version should be indicated using the
Digital Object Identifier of the specific version required (e.g.
\code{"10.5281/zenodo.3673586"}).
Defaults to \code{"latest"} such that the latest
release of the dataset with available habitat data is used.}

\item{preprocessed}{\code{logical} Should a pre-processed version of the
dataset be downloaded?
If so, the data are downloaded
from a GitHub repository
(\url{https://github.com/prioritizr/aoh/releases/tag/data}).
Otherwise, data are downloaded from original
Zenodo Digital Repository (Jung \emph{et al.} 2020b).}

\item{force}{\code{logical} should the data be downloaded even if the
the data are already available?
Defaults to \code{FALSE}.}

\item{verbose}{\code{logical} should progress be displayed while downloading
data?
Defaults to \code{TRUE}.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} object containing the level 2 habitat
fractional coverage data.
These data are available at the 1 km \eqn{\times} 1 km resolution.
Each layer corresponds to a different habitat type, and each pixel
denotes the fraction of the pixel that contains a given habitat type.
\strong{Note that pixel values are scaled to between 0 and 1000, such that
a value of 0 indicates 0\% coverage of a habitat type, and a value of
1000 indicates 100\% coverage of a habitat type.}
}
\description{
Import habitat classification data produced by Jung \emph{et al.} (2020a).
If data are not available locally, they are downloaded from
an online repository.
}
\details{
The preprocessed version of the habitat data was produced by
version was produced by projecting the data to conform
with the default template raster for processing Area of Habitat data
(see \code{\link[=get_world_behrmann_1km_rast]{get_world_behrmann_1km_rast()}}).
Additionally, data were clamped to ensure values range between
valid values (i.e. from 0 to 1000).
The preprocessed version of the habitat data can be help
run time when creating the Area of Habitat data with
the default template raster.
}
\examples{
\dontrun{
# define persistent storage location
download_dir <- rappdirs::user_data_dir("aoh")

# create download directory if needed
if (!file.exists(download_dir)) {
  dir.create(download_dir, showWarnings = FALSE, recursive = TRUE)
}

# download and import habitat data
habitat_data <- get_global_habitat_data(download_dir, version = "latest")

# preview data
print(habitat_data)

# plot data
plot(habitat_data)
}
}
\references{
Jung M, Dahal PR, Butchart SHM, Donald PF, De Lamo X, Lesiv M, Kapos V,
Rondinini C, and Visconti P (2020a) A global map of
terrestrial habitat types. \emph{Scientific Data}, 7:1--8.
Available at \url{https://doi.org/10.1038/s41597-020-00599-8}.

Jung M, Dahal PR, Butchart SHM, Donald PF, De Lamo X, Lesiv M, Kapos V,
Rondinini C, and Visconti P (2020b) A global map of
terrestrial habitat types (insert version) [Data set].
\emph{Zenodo Digital Repository}.
Available at \url{https://doi.org/10.5281/zenodo.4058819}.
}
