% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jung_lvl2_habitat_data.R
\name{get_jung_lvl2_habitat_data}
\alias{get_jung_lvl2_habitat_data}
\title{Get Jung \emph{et al.} (2020) habitat classification data (level 2)}
\usage{
get_jung_lvl2_habitat_data(
  dir = tempdir(),
  version = "latest",
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dir}{\code{character} Folder path for downloading and caching data.
By default, data are downloaded to a temporary directory (i.e.,
\code{tempdir()}). \strong{To avoid downloading the same data multiple times, it is
strongly recommended to specify a persistent storage location (see Examples
below).}}

\item{version}{\code{character} Value indicating the specific version of the
dataset
that should be downloaded. The version should be indicated using the
Digital Object Identifier of the specific version required (e.g.
\code{"10.5281/zenodo.3673586"}).
Defaults to \code{"latest"} such that the latest
release of the dataset with available habitat data is used.}

\item{force}{\code{logical} Should the data be downloaded even if the
the data are already available?
Defaults to \code{FALSE}.}

\item{verbose}{\code{logical} Should progress be displayed while downloading
data?
Defaults to \code{TRUE}.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} object containing the habitat data
(100 m resolution). Pixel values indicate the habitat classification codes.
}
\description{
Import habitat classification data (level 2) derived from Jung \emph{et al.}
(2020a).
If data are not available locally, they are downloaded from
an online repository.
}
\details{
The data were produced by obtaining the level 2 habitat
classification data from the
\href{https://doi.org/10.5281/zenodo.3666245}{Zenodo repository}
(Jung \emph{et al.} 2020b), and resampling the data (using nearest neighbor
interpolation) to the World Behrmannn coordinate reference
systems (ESRI:54017).
}
\examples{
\dontrun{
# define persistent storage location
download_dir <- rappdirs::user_data_dir("aoh")

# create download directory if needed
if (!file.exists(download_dir)) {
  dir.create(download_dir, showWarnings = FALSE, recursive = TRUE)
}

# download and import habitat data
habitat_data <- get_jung_lvl2_habitat_data(download_dir, version = "latest")

# preview data
print(habitat_data)

# plot data
plot(habitat_data)
}
}
\references{
Jung M, Dahal PR, Butchart SHM, Donald PF, De Lamo X, Lesiv M, Kapos V,
Rondinini C, and Visconti P (2020a) A global map of
terrestrial habitat types. \emph{Scientific Data}, 7:1--8.
Available at \url{https://doi.org/10.1038/s41597-020-00599-8}.

Jung M, Dahal PR, Butchart SHM, Donald PF, De Lamo X, Lesiv M, Kapos V,
Rondinini C, and Visconti P (2020b) A global map of
terrestrial habitat types (insert version) [Data set].
\emph{Zenodo Digital Repository}.
Available at \url{https://doi.org/10.5281/zenodo.4058819}.
}
\seealso{
See \code{\link[=crosswalk_jung_lvl2_data]{crosswalk_jung_lvl2_data()}} for details on which grid values correspond
to which habitat classification codes.
}
