% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jung_plvl1_habitat_data.R
\name{get_jung_plvl1_habitat_data}
\alias{get_jung_plvl1_habitat_data}
\title{Get Jung (2020) potential habitat classification data (level 1)}
\usage{
get_jung_plvl1_habitat_data(
  dir = tempdir(),
  version = "latest",
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dir}{\code{character} Folder path for downloading and caching data.
By default, data are downloaded to a temporary directory (i.e.,
\code{tempdir()}). \strong{To avoid downloading the same data multiple times, it is
strongly recommended to specify a persistent storage location (see Examples
below).}}

\item{version}{\code{character} Value indicating the specific version of the
dataset
that should be downloaded. The version should be indicated using the
Digital Object Identifier of the specific version required (e.g.
\code{"10.5281/zenodo.3673586"}).
Defaults to \code{"latest"} such that the latest
release of the dataset with available habitat data is used.}

\item{force}{\code{logical} Should the data be downloaded even if the
the data are already available?
Defaults to \code{FALSE}.}

\item{verbose}{\code{logical} Should progress be displayed while downloading
data?
Defaults to \code{TRUE}.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} object containing the potential habitat data
(100 m resolution). Pixel values indicate the potential habitat
classification codes.
}
\description{
Import potential habitat classification data derived from Jung (2020).
If data are not available locally, they are downloaded from
a Zenodo Digital Repository (https://doi.org/10.5281/zenodo.6622090).
}
\details{
The data predict the potential habitat types that would be available in a
given location if not for anthropogenic activities
(see Hengl \emph{et al.} 2018 for details).
As such, they can be used to identify locations where
restoration activities could produce certain habitat types.
This means that they can, in turn, be used to predict the location of
suitable habitat for species following the restoration activities.
Since the data do not show the historic distribution of habitat types,
they cannot be used to examine patterns of habitat loss.

The data were produced by obtaining the level 1 habitat
classification data from the
\href{https://doi.org/10.5281/zenodo.4038749}{Zenodo repository}
(Jung 2020), and resampling the data (using nearest neighbor
interpolation) to the World Behrmannn coordinate reference
systems (ESRI:54017).
}
\examples{
\dontrun{
# define persistent storage location
download_dir <- rappdirs::user_data_dir("aoh")

# create download directory if needed
if (!file.exists(download_dir)) {
  dir.create(download_dir, showWarnings = FALSE, recursive = TRUE)
}

# download and import potential habitat data
ph_data <- get_jung_plvl1_habitat_data(download_dir, version = "latest")

# preview data
print(ph_data)

# plot data
plot(ph_data)
}
}
\references{
Jung M (2020) A layer of global potential habitats (insert version)
[Data set]. \emph{Zenodo Digital Repository}.
Available at \url{https://doi.org/10.5281/zenodo.4038749}.
}
\seealso{
See \code{\link[=crosswalk_jung_lvl1_data]{crosswalk_jung_lvl1_data()}} for details on which grid values correspond
to which habitat classification codes.
}
