% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spp_summary_data.R
\name{get_spp_summary_data}
\alias{get_spp_summary_data}
\title{Get species summary data}
\usage{
get_spp_summary_data(
  x,
  dir = tempdir(),
  version = "latest",
  key = NULL,
  delay = 2,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{\code{integer} Taxon identifier for the species on the International
Union for Conservation of Nature (IUCN) Red List of Threatened
Species. For example, the taxon identifier for the species
\emph{Loxodonta africana} is \code{181008073}.}

\item{dir}{\code{character} folder path where data should be downloaded.
By default, data are downloaded to a temporary directory (i.e. \code{tempdir()}).
\strong{To avoid downloading the same data multiple times, it is strongly
recommended to specify a persistent storage location (see Examples below).}
If data are already available in folder for the specified version, then the
data are imported and not re-downloaded from the IUCN Red List.}

\item{version}{\code{character} indicating version of the IUCN Red List for
obtaining data (e.g. \code{"2021-2"}).
Although it is not possible to query past versions
of the IUCN Red List, this functionality is useful for accessing
data previously downloaded from the IUCN Red List.
Defaults to \code{"latest"} such that data are downloaded from the most recent
version of the IUCN Red List.}

\item{key}{\code{character} Token for querying the IUCN Red List API.
\strong{Please note that a token is required to download
data from the IUCN Red List.} To obtain a token, please see
\code{\link[rredlist:rl_use_iucn]{rredlist::rl_use_iucn()}}. Defaults to \code{NULL} such that the token
accessed from  the \code{"IUCN_REDLIST_KEY"} environmental variable
(which can be specified in the .Renviron file).}

\item{delay}{\code{integer} number of seconds to wait between subsequent calls
to the IUCN Red List API. Defaults to 2 seconds (as recommended by the
\pkg{rredlist} package;
\url{https://docs.ropensci.org/rredlist/articles/rredlist.html}).}

\item{force}{\code{logical} should the data be downloaded even if the
the data are already available?
Defaults to \code{FALSE}.}

\item{verbose}{\code{logical} should progress be displayed while downloading
data?
Defaults to \code{TRUE}.}
}
\value{
A table (\code{\link[tibble:tibble]{tibble::tibble()}}) object.
}
\description{
Import species summary data obtained from the
\href{https://www.iucnredlist.org/}{International Union for Conservation of Nature (IUCN) Red List of Threatened Species}.
}
\details{
Data are downloaded from the IUCN Red List using the
\code{\link[rredlist:rl_search]{rredlist::rl_search()}}. This function is essentially a wrapper
designed to help download data for multiple species and provide
caching for previously downloaded data.
}
\examples{
\dontrun{
# define species to download data for based on taxon identifiers
spp_ids <- c(12392, 22694927)

# define persistent storage location
download_dir <- rappdirs::user_data_dir("aoh")

# create download directory if needed
if (!file.exists(download_dir)) {
  dir.create(download_dir, showWarnings = FALSE, recursive = TRUE)
}

# download and import summary data
spp_summary_data <- get_spp_summary_data(spp_ids, download_dir)

# preview data (only if running R in an interactive session)
if (interactive()) {
  print(spp_summary_data)
}
}
}
\references{
Please use the following citation for data obtained from the IUCN Red List:

IUCN (insert year). IUCN Red List of Threatened Species. Version

To obtain the version number of the latest version, use
\code{\link[rredlist:rl_version]{rredlist::rl_version()}}.
}
