% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_world_berhman_1km_rast.R
\name{get_world_berhman_1km_rast}
\alias{get_world_berhman_1km_rast}
\title{World Behrman 1 km raster}
\usage{
get_world_berhman_1km_rast()
}
\value{
A \code{\link[terra:rast]{terra::rast()}} object.
}
\description{
Import a \code{\link[terra:rast]{terra::rast()}} template raster dataset delineating the
global extent at 1 km resolution using the World Behrman
coordinate reference system (ESRI:54017).
}
\details{
This raster datast is intended to be used as a template for
spatial data processing. As such, it does not contain any
meaningful data. Specifically, all grid cells contain a value of 1.
}
\examples{
# import data
x <- get_world_berhman_1km_rast()

# preview data
print(x)

# plot data
plot(x)
}
