% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersecting_ext.R
\name{intersecting_ext}
\alias{intersecting_ext}
\title{Intersecting extent}
\usage{
intersecting_ext(x, y, buffer = 0)
}
\arguments{
\item{x}{\code{\link[terra:rast]{terra::rast()}} Raster object. This object specifies the
\code{\link[terra:crs]{terra::crs()}} coordinate reference system used for the output object.}

\item{y}{\code{\link[terra:rast]{terra::rast()}} Raster object. This object specifies the
spatial extent for the output object.}

\item{buffer}{\code{numeric} buffer applied to the spatial extent of \code{y} before
reprojecting to the coordinate reference system of \code{x}. Defaults to 0.}
}
\value{
A \code{\link[terra:ext]{terra::ext()}} object.
}
\description{
Generate a \code{\link[terra:ext]{terra::ext()}} object containing the spatial extent of one
\code{\link[terra:rast]{terra::rast()}} raster object inside another \code{\link[terra:rast]{terra::rast()}} object.
}
\examples{
# define raster object
x <- rast(
  ncols = 40, nrows = 40,
  xmin = -110, xmax = -90, ymin = 40, ymax = 60,
  crs = "+proj=longlat +datum=WGS84"
)
print(x)

# define raster object with different coordinate reference system
y <- rast(
  ncols = 94, nrows = 124,
  xmin = -944881, xmax = 935118, ymin = 4664377, ymax = 7144377,
  crs = "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"
)
print(y)

# find spatial extent of x that intersects with the spatial extent of y
print(intersecting_ext(x, y))

}
\seealso{
Other geoprocessing: 
\code{\link{create_template_rast}()},
\code{\link{read_sf_n}()},
\code{\link{sf_terra_crs}()},
\code{\link{sf_terra_ext}()},
\code{\link{terra_st_bbox}()},
\code{\link{terra_st_crs}()}
}
\concept{geoprocessing}
