% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_available.R
\name{is_gdal_available}
\alias{is_gdal_available}
\title{Is GDAL available?}
\usage{
is_gdal_available()
}
\value{
A \code{logical} indicating if GDAL is available or not.
}
\description{
Check if GDAL is available for processing data.
}
\details{
The function verifies that (1) the \pkg{gdalUtils} package in installed,
(2) GDAL is installed (i.e. via \code{system("gdalinfo --version")}), and
(3) the version of GDAL installed is at least 3.0.2.
If any of these checks fail, then GDAL is not considered available.
}
\examples{
# check if GDAL is available?
print(is_gdal_available())

}
