% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_available.R
\name{is_gdal_calc_available}
\alias{is_gdal_calc_available}
\title{Is gdal_calc.py available?}
\usage{
is_gdal_calc_available()
}
\value{
A \code{logical} value indicating if it is available.
}
\description{
Check if \code{gdal_calc.py} Python script is available.
}
\details{
The \code{gdal_calc.py} Python script is used to process Area of Habitat data
when using the GDAL engine (within \code{\link[=create_spp_aoh_data]{create_spp_aoh_data()}}) .
This function determines if this script is available on the system.
On Windows systems, it first tries to access this script using the
OSGeo4W software (available at \url{https://trac.osgeo.org/osgeo4w/}).
If that fails, or if you are not using a Windows system: it then tries to
access this script using default system variables.
}
\examples{
# see if gdal_calc python script is available
print(is_gdal_calc_available())

}
