% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_available.R
\name{is_gdal_python_available}
\alias{is_gdal_python_available}
\title{Is GDAL Python interface available?}
\usage{
is_gdal_python_available()
}
\value{
A \code{logical} value indicating if it is available.
}
\description{
Check if GDAL Python scripts are available.
}
\examples{
# see if GDAL python scripts are available
print(is_gdal_python_available())

}
