% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_available.R
\name{is_grass_available}
\alias{is_grass_available}
\title{Is GRASS available?}
\usage{
is_grass_available()
}
\value{
A \code{logical} indicating if GRASS is available or not.
}
\description{
Check if GRASS is available for processing data.
}
\details{
The function verifies that GRASS can be executed from the command line
(i.e. via \code{system("grass --version")}).
}
\examples{
# check if GRASS is available?
print(is_grass_available())

}
