% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_available.R
\name{is_iucn_rl_api_available}
\alias{is_iucn_rl_api_available}
\title{Is IUCN Red List API available?}
\usage{
is_iucn_rl_api_available(key = NULL, n = 5)
}
\arguments{
\item{key}{\code{character} Token for querying the IUCN Red List API.
Defaults to \code{NULL} such that the token
accessed from  the \code{"IUCN_REDLIST_KEY"} environmental variable
(which can be specified in the .Renviron file).}

\item{n}{\code{integer} Number of times to attempt to access the API.}
}
\value{
A \code{logical} indicating if the IUCN Red List API can be accessed.
}
\description{
The International Union for Conservation of Nature (IUCN)
provides an \href{https://api.iucnredlist.org/}{API to access data from Red List of Threatened Species}.
This function checks whether data can be accessed from the API.
Please note that a token is required to access the API
(see below for instructions to obtain a token).
}
\section{Accessing the IUCN Red List API}{


You will need to obtain a token for the
\href{https://api.iucnredlist.org/}{IUCN Red List API} (if you do not have one
already). To achieve this, please visit the IUCN API website
(\url{https://api.iucnredlist.org/}), click the "Generate a token" link at the
top of the web page, and fill out the form to apply for a token.
You should then receive a token shortly after completing
the form (but not immediately). After receiving a token, please open the
\code{.Renviron} file on your computer (e.g., using \code{usethis::edit_r_environ()}).
Next, please add the following text to the file (replacing the string with
the token) and save the file:

\if{html}{\out{<div class="sourceCode">}}\preformatted{IUCN_REDLIST_KEY="your_actual_token_not_this_string"
}\if{html}{\out{</div>}}

Please restart your R session. You should now be able to access the IUCN Red
List API. To verify this, please try running the following \emph{R} code and --
assuming everything works correctly -- you should see \code{TRUE} as the output:

\if{html}{\out{<div class="sourceCode">}}\preformatted{# verify access to IUCN Red List API
is_iucn_rl_available()
}\if{html}{\out{</div>}}

If these instructions did not work, please consult the documentation for the
\pkg{rredlist} package for further details.
}

\examples{
\dontrun{
# check if IUCN Red List API is available?
is_iucn_rl_api_available()
}

}
