% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_available.R
\name{is_osgeo4w_available}
\alias{is_osgeo4w_available}
\title{Is OSGeo4W available?}
\usage{
is_osgeo4w_available()
}
\value{
A \code{logical} value indicating if it is available.
}
\description{
Check if OSGeo4W software is available.
}
\details{
This software is used to provide GDAL Python scripts on Windows systems
that are used to generate Area of Habitat data via the GDAL engine.
It can be installed from \url{https://trac.osgeo.org/osgeo4w/}.
Note that macOS and Linux systems do not require this software.
By default, it is assumed that the software is installed in the
\code{"C:/OSGeo4W"} directory. If the software is installed at a different
location, then the \code{"OSGEO4W_ROOT"} environmental variable can be
used to specify a different location.
}
\examples{
# see if OSGeo4W is available at the default location
print(is_osgeo4w_available())

\dontrun{
# specify a different location for OSGeo4W, and
# then see if OSGeo4W is available at this location
Sys.setenv("OSGEO4W_ROOT" = "C:/software/OSGeo4W")
print(is_osgeo4w_available())
}

}
