% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_habitat_codes.R
\name{iucn_habitat_codes_misc}
\alias{iucn_habitat_codes_misc}
\title{Miscellaneous IUCN Red List habitat codes}
\usage{
iucn_habitat_codes_misc()
}
\value{
A \code{character} vector of codes.
}
\description{
The \href{https://www.iucnredlist.org/}{International Union for Conservation of Nature (IUCN) Red List of Threatened Species}
provides a
\href{https://www.iucnredlist.org/resources/habitat-classification-scheme}{habitat classification scheme}
for different habitat types worldwide.
This function returns codes for habitat classes that correspond to
\emph{unknown} or \emph{other} habitats.
}
\examples{
# print codes
print(iucn_habitat_codes_misc())

}
\concept{iucn_habitat_codes}
