% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_habitat_data.R
\docType{data}
\name{iucn_habitat_data}
\alias{iucn_habitat_data}
\title{IUCN habitat classification codes}
\format{
A data frame with 126 rows and 2 columns.
}
\usage{
data(iucn_habitat_data)
}
\description{
TODO
}
\keyword{datasets}
