% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_threat_data.R
\docType{data}
\name{iucn_threat_data}
\alias{iucn_threat_data}
\title{IUCN threat classification codes}
\format{
A data frame with 130 rows and 2 columns.
Each row corresponds to a different threat class, and each
column contains information about a given threat class.
It contains columns with the following values for each threat class.
\describe{
\item{code}{A \code{character} value indicating the code of a class.}
\item{name}{A \code{character} value indicating the name of a class.}
}
}
\source{
The data were obtained from \url{https://www.iucnredlist.org/resources/threat-classification-scheme}.
}
\usage{
data(iucn_threat_data)
}
\description{
The \href{https://www.iucnredlist.org/}{International Union for Conservation of Nature (IUCN) Red List of Threatened Species}
provides a
\href{https://www.iucnredlist.org/resources/threat-classification-scheme}{threat classification scheme}
for different threatening processes.
This dataset provides details on the different IUCN threat classes.
}
\examples{
# load data
data(iucn_threat_data)

# print data
print(iucn_threat_data)
}
\keyword{datasets}
