% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spp_aoh_data.R
\name{plot_spp_aoh_data}
\alias{plot_spp_aoh_data}
\title{Plot species range and Area of Habitat data}
\usage{
plot_spp_aoh_data(x, maptype)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::st_sf()}} Object containing the species data.
This object should be produced using the \code{\link[=create_spp_aoh_data]{create_spp_aoh_data()}}
function.}

\item{maptype}{\code{character} Value indicating the name of the
the basemap to use for the plot.
See documentation for the \code{maptype} parameter in the \link[ggmap:get_stamenmap]{ggmap::get_stamenmap}
function for details.
Defaults to \code{NULL} such that no basemap is shown.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Create a map to compare species geographic range and Area of habitat data.
}
\examples{
#TODO
}
