% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_sf.R
\name{read_sf_n}
\alias{read_sf_n}
\title{Read spatial data}
\usage{
read_sf_n(dsn, layer = NULL, n = NULL)
}
\arguments{
\item{dsn}{\code{character} data source name.}

\item{layer}{\code{character} layer name. Defaults to \code{NULL}.}

\item{n}{\code{integer} number of records to import.
Defaults to \code{NULL} such that all data are imported.}
}
\value{
\code{\link[sf:sf]{sf::sf()}} object.
}
\description{
Import spatial data. If desired, only a subset of the available data
are imported.
}
\examples{
# read all data
nc <- read_sf(system.file("shape/nc.shp", package = "sf"))
print(nc)

# read only first five geometries
nc2 <- read_sf(system.file("shape/nc.shp", package = "sf"), n = 5)
print(nc2)

}
\seealso{
Other geoprocessing: 
\code{\link{create_template_rast}()},
\code{\link{sf_terra_ext}()},
\code{\link{terra_st_bbox}()},
\code{\link{terra_st_crs}()}
}
\concept{geoprocessing}
