% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spp_range_data.R
\name{read_spp_range_data}
\alias{read_spp_range_data}
\title{Read species range data}
\usage{
read_spp_range_data(path, n = NULL)
}
\arguments{
\item{path}{\code{character} File path to the data (zip archive) file.}

\item{n}{\code{numeric} Number of features in the dataset to import.
Defaults to \code{NULL} such that all available data is imported.}
}
\value{
A \code{\link[sf:sf]{sf::sf()}} object containing the dataset.
}
\description{
Import species geographic range (i.e. extent of occurrence) data obtained
from the International Union for Conservation of Nature (IUCN) Red List of
Threatened Species (\url{https://www.iucnredlist.org/}).
}
\details{
Data for amphibians, reptiles, and mammals can be obtained directly from
the International Union for Conservation of Nature (IUCN) Red List website
(see https://www.iucnredlist.org/resources/spatial-data-download).
Data for birds can be obtained by requesting data from
\href{http://www.birdlife.org/}{BirdLife International}
(see http://datazone.birdlife.org/species/requestdis).
}
\examples{
# find file path for simulated data following the IUCN Red List format
path <- system.file("XENOMORPHS_TERRESTRIAL_ONLY.zip")

# import data
sim_spp_range_data <- read_spp_range_data(path)

# preview data (only if running R in an interactive session)
if (interactive()) {
  print(sim_spp_range_data)
}

# plot data (only if running R in an interactive session)
if (interactive()) {
  print(sim_spp_range_data)
}
}
