% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_sf.R
\name{sf_terra_crs}
\alias{sf_terra_crs}
\title{Coordinate reference system}
\usage{
sf_terra_crs(x)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::st_sf()}} Object.}
}
\value{
A \code{character} value.
}
\description{
Create a \code{character} object to describe the coordinate reference system
of a \code{\link[sf:sf]{sf::st_sf()}} object following the \pkg{terra} package conventions.
}
\examples{
# import sf object
nc <- read_sf(system.file("shape/nc.shp", package = "sf"))

# create terra CRS value
sf_terra_crs(nc)

}
\seealso{
Other geoprocessing: 
\code{\link{create_template_rast}()},
\code{\link{read_sf_n}()},
\code{\link{sf_terra_ext}()},
\code{\link{terra_st_bbox}()},
\code{\link{terra_st_crs}()}
}
\concept{geoprocessing}
