% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_sf.R
\name{sf_terra_ext}
\alias{sf_terra_ext}
\title{Spatial extent}
\usage{
sf_terra_ext(x)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::st_sf()}} Spatial object.}
}
\value{
A \code{\link[terra:ext]{terra::ext()}} extent object.
}
\description{
Create a \code{\link[terra:ext]{terra::ext()}} object to describe the spatial extent of
a \code{\link[sf:sf]{sf::st_sf()}} object following the \pkg{terra} package conventions.
}
\examples{
# import sf object
nc <- read_sf(system.file("shape/nc.shp", package = "sf"))

# create terra extent object (i.e. SpatExtent)
sf_terra_ext(nc)

}
\seealso{
Other geoprocessing: 
\code{\link{create_template_rast}()},
\code{\link{intersecting_ext}()},
\code{\link{read_sf_n}()},
\code{\link{sf_terra_crs}()},
\code{\link{terra_st_bbox}()},
\code{\link{terra_st_crs}()}
}
\concept{geoprocessing}
