% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_available.R
\name{system_gdal_version}
\alias{system_gdal_version}
\title{System GDAL version}
\usage{
system_gdal_version()
}
\value{
A \code{character} value describing the version of GDAL installed.
If GDAL is not installed, then a missing (\code{NA}) value is returned.
}
\description{
Find the version of the Geospatial Data Abstraction Library (GDAL)
that is currently installed on the system.
}
\examples{
# show version of GDAL installed
print(system_gdal_version())

}
