% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_terra.R
\name{terra_combine}
\alias{terra_combine}
\title{Combine}
\usage{
terra_combine(x)
}
\arguments{
\item{x}{\code{list} of \code{\link[terra:rast]{terra::rast()}} objects.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} object.
}
\description{
Align and combine a list of multiple \code{\link[terra:rast]{terra::rast()}} objects into a
single \code{\link[terra:rast]{terra::rast()}} object with multiple layers.
}
\examples{
# create a raster
x <- rast(
  ncols = 10, nrows = 10,
  xmin = 0, xmax = 10, ymin = 0, ymax = 10,
  crs = "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"
)
print(x)

# create another raster
y <- rast(
  ncols = 5, nrows = 5,
  xmin = 90, xmax = 95, ymin = 80, ymax = 85,
  crs = "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"
)
print(y)

# combine them together
z <- terra_combine(list(x, y))

# plot result
plot(z)

}
