% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_terra.R
\name{terra_fasterize}
\alias{terra_fasterize}
\title{Fasterize}
\usage{
terra_fasterize(sf, raster, ...)
}
\arguments{
\item{sf}{\code{\link[sf:sf]{sf::st_sf()}} Object.}

\item{raster}{\code{\link[terra:rast]{terra::rast()}} Object.}

\item{...}{Additional arguments passed to \code{\link[fasterize:fasterize]{fasterize::fasterize()}}.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} object.
}
\description{
This function converts a \code{\link[sf:st_as_sf]{sf::st_as_sf()}} object to a \code{\link[terra:rast]{terra::rast()}}
object using \code{\link[fasterize:fasterize]{fasterize::fasterize()}}. It is similar to
\code{\link[terra:rasterize]{terra::rasterize()}}, except that it has greater performance.
}
\examples{
\dontrun{
# load data
nc <- read_sf(system.file("shape/nc.shp", package = "sf"))
nc <- st_transform(nc, st_crs("ESRI:54017"))
r <- get_world_behrmann_1km_rast()

# convert to raster
nc_rast <- terra_fasterize(nc, r)

# plot result
plot(nc_rast)
}
}
