% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_terra.R
\name{terra_force_disk}
\alias{terra_force_disk}
\title{Force file-backed raster}
\usage{
terra_force_disk(x, filename = tempfile(fileext = ".tif"), ...)
}
\arguments{
\item{x}{\code{\link[terra:rast]{terra::rast()}} Raster object.}

\item{filename}{\code{character} File path to store data.
Defaults to a temporarily (GeoTIFF) file.}

\item{...}{Arguments passed to \code{\link[terra:writeRaster]{terra::writeRaster()}} for saving data.}
}
\value{
A `\code{\link[terra:rast]{terra::rast()}} raster object.
}
\description{
Force a raster (\code{\link[terra:rast]{terra::rast()}}) object to have data stored on disk.
}
