% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra_gdal_crop.R
\name{terra_gdal_crop}
\alias{terra_gdal_crop}
\title{Crop a raster using GDAL}
\usage{
terra_gdal_crop(
  x,
  ext,
  n_threads = 1,
  filename = tempfile(fileext = ".tif"),
  datatype = "FLT4S",
  tiled = FALSE,
  bigtiff = FALSE,
  output_raster = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{\code{\link[terra:rast]{terra::rast()}} Raster object with source data.}

\item{ext}{\code{\link[terra:ext]{terra::ext()}} Raster extent object.}

\item{n_threads}{\code{integer} Number of computational threads to use
for data processing.
To reduce run time, it is strongly recommended to set this
parameter based on available resources (see Examples section below).
Defaults to 1.}

\item{filename}{\code{character} Filename for output raster.
Defaults to \code{tempfile(fileext = ".tif")}.}

\item{datatype}{\code{character} Value indicating the data type for saving data.
Defaults to \code{"FLT4S"}.}

\item{tiled}{\code{logical} Value indicating if GeoTIFF files should be tiled.
Defaults to \code{FALSE}.}

\item{bigtiff}{\code{logical} Value indicating the data should be stored in
BIGTIFF format.
Defaults to \code{FALSE}.}

\item{output_raster}{\code{logical} Should a raster (\code{\link[terra:rast]{terra::rast()}})
object be returned?
If \code{FALSE} then the file path for the resulting file is returned.
Defaults to \code{TRUE}.}

\item{verbose}{\code{logical} Should information be displayed during processing?
Defaults to \code{TRUE}.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} raster object.
}
\description{
This function is a wrapper for \code{\link[gdalUtils:gdal_translate]{gdalUtils::gdal_translate()}}.
}
\examples{
# please ensure that the gdalUtils package is installed and
# GDAL system binaries are installed to run this example

\dontshow{if (is_gdal_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# create raster with data
x <- rast(
  ncols = 40, nrows = 40, xmin = -110, xmax = -90, ymin = 40, ymax=60,
  crs = "+proj=longlat +datum=WGS84"
)
values(x) <- seq_len(ncell(x))

# create extent for cropping
y <- ext(x) - c(5, 2.5, 1, 1.5)

# crop data
z <- terra_gdal_crop(x, y)

# preview result
print(z)
\dontshow{\}) # examplesIf}
}
