% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra_gdal_project.R
\name{terra_gdal_project}
\alias{terra_gdal_project}
\title{Project a raster using GDAL}
\usage{
terra_gdal_project(
  x,
  y,
  method = "bilinear",
  n_threads = 1,
  filename = tempfile(fileext = ".tif"),
  datatype = "FLT4S",
  cache_limit = 200,
  tiled = FALSE,
  bigtiff = FALSE,
  compress = "LZW",
  verbose = TRUE,
  NAflag = NULL,
  output_raster = TRUE
)
}
\arguments{
\item{x}{\code{\link[terra:rast]{terra::rast()}} Raster object with source data.}

\item{y}{\code{\link[terra:rast]{terra::rast()}} Raster object specifying spatial properties for
output data.}

\item{method}{\code{character} Name of interpolation method.}

\item{n_threads}{\code{integer} Number of computational threads to use
for data processing.
To reduce run time, it is strongly recommended to set this
parameter based on available resources (see Examples section below).
Defaults to 1.}

\item{filename}{\code{character} Filename for output raster.
Defaults to \code{tempfile(fileext = ".tif")}.}

\item{datatype}{\code{character} Value indicating the data type for saving data.
Defaults to \code{"FLT4S"}.}

\item{cache_limit}{\code{integer} Number of MB to use for GDAL caching.
Defaults to 200.}

\item{tiled}{\code{logical} Value indicating if GeoTIFF files should be tiled.
Defaults to \code{FALSE}.}

\item{bigtiff}{\code{logical} Value indicating the data should be stored in
BIGTIFF format.
Defaults to \code{FALSE}.}

\item{compress}{\code{character} Value indicating compression format.
Available options include \code{"LZW"} and \code{"DEFLATE"}.
Defaults to \code{"LZW"}.}

\item{verbose}{\code{logical} Should information be displayed during processing?
Defaults to \code{TRUE}.}

\item{NAflag}{\code{numeric} Value for representing missing (\code{NA}) values.
A \code{"none"} value can also be used to indicate that no flag should be set.
Defaults to \code{NULL} such that the value is determined automatically.}

\item{output_raster}{\code{logical} Should a raster (\code{\link[terra:rast]{terra::rast()}})
object be returned?
If \code{FALSE} then the file path for the resulting file is returned.
Defaults to \code{TRUE}.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} raster object.
}
\description{
This function is a wrapper for \code{\link[gdalUtilities:gdalwarp]{gdalUtilities::gdalwarp()}} for use with
\pkg{terra} objects.
}
\examples{
# please ensure that the gdalUtilities package is installed
# to run this example

\dontshow{if (requireNamespace("gdalUtilities", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# create raster with data
x <- rast(
  ncols = 40, nrows = 40, xmin = -110, xmax = -90, ymin = 40, ymax=60,
  crs = "+proj=longlat +datum=WGS84"
)
values(x) <- seq_len(ncell(x))

# create raster to define spatial properties for projection
y <- rast(
  ncols = 94, nrows = 124, xmin = -944881, xmax = 935118, ymin = 4664377,
  ymax = 7144377,
  crs = "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"
)

# project data
z <- terra_gdal_project(x, y)

# preview result
print(z)
\dontshow{\}) # examplesIf}
}
