% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra_gdal_rasterize.R
\name{terra_gdal_rasterize}
\alias{terra_gdal_rasterize}
\title{Convert vector data into a raster using GDAL}
\usage{
terra_gdal_rasterize(
  x,
  sf,
  burn = 1,
  invert = FALSE,
  update = FALSE,
  n_threads = 1,
  filename = tempfile(fileext = ".tif"),
  sf_filename = tempfile(fileext = ".gpkg"),
  datatype = "FLT4S",
  tiled = FALSE,
  bigtiff = FALSE,
  compress = "LZW",
  verbose = TRUE,
  output_raster = TRUE
)
}
\arguments{
\item{x}{\code{\link[terra:rast]{terra::rast()}} Raster object with source data.}

\item{sf}{\code{\link[sf:sf]{sf::st_sf()}} Spatial object to rasterize.}

\item{burn}{\code{numeric} Value for encoding the vector data.
Defaults to 1.}

\item{invert}{\code{logical} Should the burn process be inverted?
Defaults to \code{FALSE}.}

\item{update}{\code{logical} Should the result by producing by updating
the argument to \code{x}?
If \code{FALSE} then the argument to \code{x} is only used to specify the
spatial properties of the resulting raster
(i.e. values have on the result),
Defaults to \code{FALSE}.}

\item{n_threads}{\code{integer} Number of computational threads to use
for data processing.
To reduce run time, it is strongly recommended to set this
parameter based on available resources (see Examples section below).
Defaults to 1.}

\item{filename}{\code{character} Filename for output raster.
Defaults to \code{tempfile(fileext = ".tif")}.}

\item{sf_filename}{\code{character} File name to temporarily save argument
to \code{sf}.
Defaults to a temporary (geopackage) file.}

\item{datatype}{\code{character} Value indicating the data type for saving data.
Defaults to \code{"FLT4S"}.}

\item{tiled}{\code{logical} Value indicating if GeoTIFF files should be tiled.
Defaults to \code{FALSE}.}

\item{bigtiff}{\code{logical} Value indicating the data should be stored in
BIGTIFF format.
Defaults to \code{FALSE}.}

\item{compress}{\code{character} Value indicating compression format.
Defaults to \code{"LZW"}.}

\item{verbose}{\code{logical} Should information be displayed during processing?
Defaults to \code{TRUE}.}

\item{output_raster}{\code{logical} Should a raster (\code{\link[terra:rast]{terra::rast()}})
object be returned?
If \code{FALSE} then the file path for the resulting file is returned.
Defaults to \code{TRUE}.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} raster object.
}
\description{
This function is a wrapper for \code{\link[gdalUtils:gdal_rasterize]{gdalUtils::gdal_rasterize()}}.
}
\examples{
# please ensure that the gdalUtils package is installed and
# GDAL system binaries are installed to run this example

\dontshow{if (is_gdal_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# TODO
\dontshow{\}) # examplesIf}
}
