% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_terra.R
\name{terra_on_disk}
\alias{terra_on_disk}
\title{On disk?}
\usage{
terra_on_disk(x)
}
\arguments{
\item{x}{\code{\link[terra:rast]{terra::rast()}} Raster object.}
}
\value{
A \code{logical} indicating if the data is stored on disk.
}
\description{
Check if a raster (\code{\link[terra:rast]{terra::rast()}}) object is stored on disk?
}
\details{
The data is only considered available on disk if the data
is not stored in memory and is only obtained from a single file source.
}
