% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_terra.R
\name{terra_st_bbox}
\alias{terra_st_bbox}
\title{Bounding box}
\usage{
terra_st_bbox(x)
}
\arguments{
\item{x}{\code{\link[terra:rast]{terra::rast()}} Raster object.}
}
\value{
A sf::st_crs()] object.
}
\description{
Create a \code{\link[sf:st_bbox]{sf::st_bbox()}} object to describe the spatial extent
of a \code{\link[terra:rast]{terra::rast()}} object following the \pkg{sf} package
conventions.
}
\examples{
# create raster object
x <- rast(
  ncols = 40, nrows = 40,
  xmin = -110, xmax = -90, ymin = 40, ymax = 60,
  crs = "+proj=longlat +datum=WGS84"
)

# extract bounding box
terra_st_bbox(x)

}
\seealso{
Other geoprocessing: 
\code{\link{create_template_rast}()},
\code{\link{read_sf_n}()},
\code{\link{sf_terra_crs}()},
\code{\link{sf_terra_ext}()},
\code{\link{terra_st_crs}()}
}
\concept{geoprocessing}
