% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_terra.R
\name{terra_temp_path}
\alias{terra_temp_path}
\title{Temporary raster file path}
\usage{
terra_temp_path(fileext = ".tif")
}
\arguments{
\item{fileext}{\code{character} Defaults to ".tif".}
}
\value{
A \code{character} file path.
}
\description{
Create a temporary raster file path using the \pkg{terra} package
to ensure cross-platform compatibility.
}
\details{
This function is used to create temporary raster files that works
cross-platform. This is especially important because the \code{\link[=tempfile]{tempfile()}}
function will sometimes replace characters in the file path with a \code{"~"}
symbol -- resulting in invalid file paths -- if the full file path is
too long.
}
