% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_spp_range_data.R
\name{clean_spp_range_data}
\alias{clean_spp_range_data}
\title{Clean species range data}
\usage{
clean_spp_range_data(
  x,
  crs = sf::st_crs("ESRI:54017"),
  snap_tolerance = 1,
  geometry_precision = 1500
)
}
\arguments{
\item{x}{A \code{\link[sf:sf]{sf::sf()}} object containing species geographic range data.}

\item{crs}{\verb{character or }integer` object representing a coordinate
reference system. Defaults to World Behrmann (ESRI:54017).}

\item{snap_tolerance}{\code{numeric} tolerance for snapping geometry to a grid
for resolving invalid geometries. Defaults to 1 meter.}

\item{geometry_precision}{\code{numeric} level of precision for processing the
spatial data (used with \code{\link[sf:st_precision]{sf::st_set_precision()}}). The default is
1500 (higher values indicate higher precision). This level of precision is
generally suitable for analyses at the national-scale. For analyses at
finer-scale resolutions, consider using a greater value (e.g.
10000).}
}
\value{
A \code{\link[sf:sf]{sf::sf()}} object containing the dataset.
}
\description{
Prepare species geographic range (i.e. extent of occurrence) data obtained
from the International Union for Conservation of Nature (IUCN) Red List of
Threatened Species (\url{https://www.iucnredlist.org/}) for analysis.
}
\details{
This function applies the following data cleaning procedures:

\enumerate{

\item Column names are standardized. This involves converting them to lower
case characters and fixing any spelling mistakes.

\item Places where the species' presence is not "known or thought very likely
to occur currently" are excluded (i.e. filtering based on
\code{presence == 1}).

\item Places where the species' presence is not "native to the area"
are excluded (i.e. filtering based on \code{origin == 1}).

\item Places where the species' seasonal occurrence is uncertain are excluded
(i.e. filtering based on \code{seasonal != 5}).

\item Species that are not terrestrial are excluded (i.e. filtering based on
where \code{terrestrial == "true"}, \code{freshwater == "false"}, and
\code{marine == "false"}).

\item Fix any potential geometry issues (using \code{\link[sf:valid]{sf::st_make_valid()}}).

\item Wrap geometries to date line (using \code{\link[sf:st_transform]{sf::st_wrap_dateline()}}).

\item Fix any potential geometry issues (i.e. using \code{\link[sf:valid]{sf::st_make_valid()}}).

\item Reproject data to specified coordinate system
(i.e. using \code{\link[sf:st_transform]{sf::st_transform()}}).

\item Fix any potential geometry issues (i.e. using \code{\link[sf:valid]{sf::st_make_valid()}}).

\item Snap geometries to spatial grid (using \code{\link[lwgeom:st_snap_to_grid]{lwgeom::st_snap_to_grid()}}).
Note that this step is only performed if the argument to \code{snap_tolerance}
is greater than zero.

\item Fix any potential geometry issues (using \code{\link[sf:valid]{sf::st_make_valid()}}).

\item Data are spatially dissolved so that each taxon is represented by
a separate geometry.

\item Fix any potential geometry issues (i.e. using \code{\link[sf:valid]{sf::st_make_valid()}}).

}

For more information on criteria used in steps 2 -- 4, please refer to
IUCN SSC Red List Technical Working Group (2021).
}
\examples{
\dontrun{
# find file path for simulated data following the IUCN Red List format
path <- system.file("XENOMORPHS_TERRESTRIAL_ONLY.zip")

# import data
sim_spp_range_data <- read_spp_range_data(path)

# clean data
sim_spp_range_data <- clean_spp_range_data(sim_spp_range_data)

# preview data (only if running R in an interactive session)
if (interactive()) {
  print(sim_spp_range_data)
}

# plot data (only if running R in an interactive session)
if (interactive()) {
  print(sim_spp_range_data)
}
}
}
\references{
IUCN SSC Red List Technical Working Group (2021)
Mapping Standards and Data Quality for the IUCN Red List Spatial Data
(Version 1.19). International Union for Conservation of Nature (IUCN).
\url{https://nc.iucnredlist.org/redlist/resources/files/1539098236-Mapping_Standards_Version_1.16_2018.pdf}
}
