% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aoh_data.R
\name{create_aoh_data}
\alias{create_aoh_data}
\title{Create Area of Habitat data}
\usage{
create_aoh_data(
  x,
  output_dir,
  cache_dir = tempdir(),
  spp_summary_data = NULL,
  spp_habitat_data = NULL,
  elevation_data = NULL,
  habitat_data = NULL,
  template_data = NULL,
  iucn_version = "latest",
  habitat_version = "latest",
  key = NULL,
  force = FALSE,
  parallel_n_threads = 1,
  parallel_cluster = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::sf()}} Spatial data delineating species geographic ranges
obtained from the IUCN Red List (\url{https://www.iucnredlist.org/}).}

\item{output_dir}{\code{character} folder path to save raster files
(GeoTIFF format) containing the Area of Habitat data.}

\item{cache_dir}{\code{character} folder path for downloading and caching data.
By default, a temporary directory is used (i.e. \code{tempdir()}).
\strong{To avoid downloading the same data multiple times, it is strongly
recommended to specify a persistent storage location (see Examples below).}}

\item{spp_summary_data}{\code{\link[tibble:tibble]{tibble::tibble()}} Table containing summary
information for each species (in the argument to \code{x}).
Specifically, the argument should contain the following columns: \code{"id_no"},
\code{"elevation_lower"}, and \code{"elevation_upper"} columns.
Here, \code{"id_no"} corresponds to the species' taxon identifier
(also present in \code{x}), and the \code{"elevation_lower"} and \code{"elevation_upper"}
columns indicate the lowest and highest elevations that contain habitat
for the species.
Defaults to \code{NULL} such that data are automatically obtained from the
latest version of the IUCN Red List (\url{https://www.iucnredlist.org}).}

\item{spp_habitat_data}{\code{\link[tibble:tibble]{tibble::tibble()}} Table containing habitat
preference information for each species (in the argument to \code{x}).
Specifically, the argument should contain the following columns: \code{"id_no"},
\code{"code"}, \code{"suitability"}, \code{"season"} columns.
Here, \code{"id_no"} corresponds to the species' taxon identifier
(also present in \code{x}), \code{"code"} indicates a habitat classification code
that is suitable for the species (i.e. based on layer names in the
argument to habitat_data), \code{"suitability"} indicates the level suitability
of the habitat class for a given species (e.g. using values such
as \code{"Suitable"} or \code{"Marginal"}), and \code{"season"} indicates
if the habitat class is only suitable for a particular seasonal
distribution (e.g. \code{"Breeding"}).
Defaults to \code{NULL} such that data are automatically obtained from the
latest version of the IUCN Red List (\url{https://www.iucnredlist.org}).}

\item{elevation_data}{\code{\link[terra:rast]{terra::rast()}} Raster data delineating the
worldwide elevation data (e.g. Jung \emph{et al.} 2020a).
Defaults to \code{NULL} such that data
are automatically obtained from Jung \emph{et al.} 2020b
(using \code{\link[=get_global_elevation_data]{get_global_elevation_data()}}).}

\item{habitat_data}{\code{\link[terra:rast]{terra::rast()}} Multi-layer raster data delineating the
coverage of different habitat classes across world
(e.g. Jung \emph{et al.} 2020a).
Each layer should correspond to a different habitat class,
and the name of each layer should contain a unique code used to identify
which places contain suitable habitat for a given species
(per the \code{"code"} column in the the argument to \code{spp_habitat_data}).
Defaults to \code{NULL} such that data
are automatically obtained from Jung \emph{et al.} 2020b
(using \code{\link[=get_global_habitat_data]{get_global_habitat_data()}}).}

\item{template_data}{\code{\link[terra:rast]{terra::rast()}} Raster specify the resolution,
coordinate reference system, and dimensionality for the output raster files.
Defaults to \code{NULL} such that a template raster dataset is automatically
(i.e. argument to \code{x}) assuming a 1 km \eqn{\times} 1 km resolution.}

\item{iucn_version}{\code{character} Version of the
IUCN Red List dataset that should be used. See documentation for the
the \code{version} parameter in the \code{\link[=get_spp_summary_data]{get_spp_summary_data()}} function
for further details.
Defaults to \code{"latest"} such that the most recent version of the dataset is
used.}

\item{habitat_version}{\code{character} Version of the
habitat dataset that should be used. See documentation for the
the \code{version} parameter in the \code{\link[=get_global_habitat_data]{get_global_habitat_data()}} function
for further details.
Defaults to \code{"latest"} such that the most recent version of the dataset is
used.}

\item{key}{\code{character} Token for querying the IUCN Red List API.
\strong{Please note that a token is required to download
data from the IUCN Red List.} To obtain a token, please see
\code{\link[rredlist:rl_use_iucn]{rredlist::rl_use_iucn()}}. Defaults to \code{NULL} such that the token
accessed from  the \code{"IUCN_REDLIST_KEY"} environmental variable
(which can be specified in the .Renviron file).}

\item{force}{\code{logical} should the data be downloaded even if the
the data are already available?
Defaults to \code{FALSE}.}

\item{parallel_n_threads}{\code{integer} Number of computational threads to use
for data processing.
\strong{To reduce run time, it is strongly recommended to set this
parameter based on the number of available threads (see Examples below).}
Default to 1.}

\item{parallel_cluster}{\code{character} Name of cluster method for
processing data in parallel. Available options are \code{"FORK"} and \code{"PSOCK"}.
Defaults to \code{NULL} such that \code{"FORK"} is used on Unix operating systems,
and \code{"PSOCK"} otherwise.}

\item{verbose}{\code{logical} Should progress be displayed while downloading
and processing data?
Defaults to \code{TRUE}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame containing the metadata for the
the Area of Habitat data. This object contains the same columns as the
argument to \code{x} along with an additional \code{path} column that contains
the file path for each Area of Habitat (GeoTIFF) raster file.
Since Area of Habitat data cannot be produced for species lacking
habitat preference data, species that do lack such data
(per arguments to \code{spp_habitat_data} and \code{habitat_data})
have missing (\code{NA}) values for the \code{path} column.
}
\description{
Create Area of Habitat (AOH) data for species based on their altitudinal and
habitat preferences (Brooks \emph{et al.} 2019).
}
\examples{
\dontrun{
# load built in datasets
data(sim_spp_summary_data, package = "aoh")
data(sim_spp_habitat_data, package = "aoh")

# find file path for simulated data following the IUCN Red List format
path <- system.file("XENOMORPHS_TERRESTRIAL_ONLY.zip")

# import species range data
sim_spp_range_data <- read_spp_range_data(path)

# specify settings for data processing
output_dir <- tempdir()                  # folder to save AOH data
cache_dir <- rappdirs::user_data_dir()   # persistent storage location
n_threads <- parallel::detectCores() - 1 # recommended for best performance

# create Area of Habitat data for species
sim_aoh_data <- create_aoh_data(
  x = sim_spp_range_data,
  output_dir = output_dir, ## change this
  spp_summary_data = sim_spp_summary_data,
  spp_habitat_data = sim_spp_habitat_data,
  parallel_n_threads = n_threads,
  cache_dir = cache_dir
)

# preview result
print(sim_aoh_data)

# import AOH data
aoh_data <- lapply(sim_aoh_data, terra::rast)

# compare range data (EOO) and AOH data for first species
#TODO
}
}
\references{
Amatulli G, Domisch S, Tuanmu M-N, Parmentier B, Ranipeta A, Malczyk J, and
Jetz W (2018) A suite of global, cross-scale topographic variables for
environmental and biodiversity modeling. Scientific Data, 5:180040.
\url{https://doi.org/10.1038/sdata.2018.40}

Brooks TM, Pimm SL, Akçakaya HR, Buchanan GM, Butchart SHM, Foden W,
Hilton-Taylor C, Hoffmann M, Jenkins CN, Joppa L, Li BV, Menon V,
Ocampo-Peñuela N, Rondinini C (2019) Measuring terrestrial Area of Habitat
(AOH) and its Utility for the IUCN Red List. Trends in Ecology & Evolution.
34:977--986. \url{doi:10.1016/j.tree.2019.06.009}

Jung M, Dahal PR, Butchart SH, Donald PF, De Lamo X, Lesiv M, Kapos V,
Rondinini C, and Visconti P (2020a) A global map of
terrestrial habitat types. Scientific data, 7:1--8.
\url{https://doi.org/10.1038/s41597-020-00599-8}

Jung M, Dahal PR, Butchart SH, Donald PF, De Lamo X, Lesiv M, Kapos V,
Rondinini C, and Visconti P (2020b) A global map of
terrestrial habitat types (insert version) [Data set]. Zenodo.
\url{https://doi.org/10.5281/zenodo.4058819}
}
