% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spp_aoh_data.R
\name{create_spp_aoh_data}
\alias{create_spp_aoh_data}
\title{Create Area of Habitat data}
\usage{
create_spp_aoh_data(
  x,
  output_dir,
  spp_summary_data = NULL,
  spp_habitat_data = NULL,
  elevation_data = NULL,
  habitat_data = NULL,
  template_data = get_world_behrmann_1km_rast(),
  cache_dir = tempdir(),
  iucn_version = "latest",
  habitat_version = "latest",
  key = NULL,
  force = FALSE,
  n_threads = 1,
  use_gdal = is_gdal_available(),
  omit_habitat_codes = iucn_habitat_codes_marine(),
  verbose = TRUE
)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::sf()}} Spatial data delineating species geographic ranges
obtained from the \href{https://www.iucnredlist.org/}{IUCN Red List}.
See below for details.}

\item{output_dir}{\code{character} folder path to save raster files
(GeoTIFF format) containing the Area of Habitat data.}

\item{spp_summary_data}{\code{\link[tibble:tibble]{tibble::tibble()}} Table containing summary
information for each species (in the argument to \code{x}).
Specifically, the argument should contain the following columns: \code{"id_no"},
\code{"elevation_lower"}, and \code{"elevation_upper"} columns.
Here, \code{"id_no"} corresponds to the species' taxon identifier
(also present in \code{x}), and the \code{"elevation_lower"} and \code{"elevation_upper"}
columns indicate the lowest and highest elevations that contain habitat
for the species.
Defaults to \code{NULL} such that data are automatically obtained from the
latest version of the \href{https://www.iucnredlist.org}{IUCN Red List}.}

\item{spp_habitat_data}{\code{\link[tibble:tibble]{tibble::tibble()}} Table containing habitat
preference information for each species (in the argument to \code{x}).
Specifically, the argument should contain the following columns: \code{"id_no"},
\code{"code"}, \code{"suitability"}, \code{"season"} columns.
Here, \code{"id_no"} corresponds to the species' taxon identifier
(also present in \code{x}), \code{"code"} indicates a habitat classification code
that is suitable for the species (i.e. based on layer names in the
argument to habitat_data), \code{"suitability"} indicates the level suitability
of the habitat class for a given species (e.g. using values such
as \code{"Suitable"} or \code{"Marginal"}), and \code{"season"} indicates
if the habitat class is only suitable for a particular seasonal
distribution (e.g. \code{"Breeding"}).
Defaults to \code{NULL} such that data are automatically obtained from the
latest version of the \href{https://www.iucnredlist.org}{IUCN Red List}.}

\item{elevation_data}{\code{\link[terra:rast]{terra::rast()}} Raster data delineating the
worldwide elevation data (e.g. Amatulli \emph{et al.} 2018).
Defaults to \code{NULL} such that data
are automatically obtained (using \code{\link[=get_global_elevation_data]{get_global_elevation_data()}}).
If the data are obtained automatically, then a preprocessed version
of the habitat data will be used to reduce processing time.}

\item{habitat_data}{\code{\link[terra:rast]{terra::rast()}} Multi-layer raster data delineating the
coverage of different habitat classes across world
(e.g. Jung \emph{et al.} 2020a,b).
Each layer should correspond to a different habitat class,
and the name of each layer should contain a unique code used to identify
which places contain suitable habitat for a given species
(per the \code{"code"} column in the the argument to \code{spp_habitat_data}).
For each layer, pixel values should indicate the percentage of available
an available habitat class.
\strong{Critically, a value of 1000 should indicate 100\%, and a value of 0
should indicate 0\%.}
Defaults to \code{NULL} such that data
are automatically obtained (using \code{\link[=get_global_habitat_data]{get_global_habitat_data()}}).
If the data are obtained automatically, then a preprocessed version
of the habitat data will be used to reduce processing time.}

\item{template_data}{\code{\link[terra:rast]{terra::rast()}} Template raster specifying the
resolution, coordinate reference system, and dimensionality for the
output raster files.
Defaults to \code{\link[=get_world_behrmann_1km_rast]{get_world_behrmann_1km_rast()}} which is a global 1 km
resolution raster under the World Behrmann coordinate reference system
(ESRI:54017).}

\item{cache_dir}{\code{character} folder path for downloading and caching data.
By default, a temporary directory is used (i.e. \code{tempdir()}).
\strong{To avoid downloading the same data multiple times, it is strongly
recommended to specify a persistent storage location (see Examples below).}}

\item{iucn_version}{\code{character} Version of the
IUCN Red List dataset that should be used. See documentation for the
the \code{version} parameter in the \code{\link[=get_spp_summary_data]{get_spp_summary_data()}} function
for further details.
Defaults to \code{"latest"} such that the most recent version of the dataset is
used.}

\item{habitat_version}{\code{character} Version of the
habitat dataset that should be used. See documentation for the
the \code{version} parameter in the \code{\link[=get_global_habitat_data]{get_global_habitat_data()}} function
for further details.
This parameter is only used if habitat data are obtained
automatically (i.e. the argument to \code{habitat_data} is \code{NULL}).
Defaults to \code{"latest"} such that the most recent version of the dataset is
used if data need to be obtained.}

\item{key}{\code{character} Token for querying the IUCN Red List API.
Defaults to \code{NULL} such that the token
accessed from  the \code{"IUCN_REDLIST_KEY"} environmental variable
(which can be specified in the .Renviron file).}

\item{force}{\code{logical} should the data be downloaded even if the
the data are already available?
Defaults to \code{FALSE}.}

\item{n_threads}{\code{integer} Number of computational threads to use
for data processing.
To reduce run time, it is strongly recommended to set this
parameter based on available resources (see Examples section below).
Note that parallel processing is only used for processing the
habitat classification and elevation data.
As such, this parameter will have no influence when using preprocessed
datasets.
Defaults to 1.}

\item{use_gdal}{\code{logical} indicating if GDAL should be used for
projecting and cropping raster data?
If GDAL is installed on the system and the \pkg{gdalUtils} package
is installed, this can dramatically reduce run time.
Defaults to \code{TRUE} if GDAL is available (per \code{\link[=is_gdal_available]{is_gdal_available()}}).}

\item{omit_habitat_codes}{\code{character} Habitat classification codes
to omit from resulting Area of Habitat data.
Please see the \href{https://www.iucnredlist.org/resources/habitat-classification-scheme}{IUCN Red List Habitat Classification Scheme}
for the full range of habitat classification codes.
For example,
if the aim is to identify natural places that contain suitable conditions,
then areas classified as anthropogenically modified
(\code{\link[=iucn_habitat_codes_artificial]{iucn_habitat_codes_artificial()}}),
introduced vegetation (\code{\link[=iucn_habitat_codes_introduced]{iucn_habitat_codes_introduced()}},
or unknown habitat (\code{\link[=iucn_habitat_codes_misc]{iucn_habitat_codes_misc()}}) should
be excluded.
Defaults to \code{\link[=iucn_habitat_codes_marine]{iucn_habitat_codes_marine()}}, such that marine
habitats are excluded.}

\item{verbose}{\code{logical} Should progress be displayed while processing data?
Defaults to \code{TRUE}.}
}
\value{
A \code{\link[sf:sf]{sf::st_sf()}} object containing range maps for the species
distributions used to generate the Area of Habitat data and additional
columns describing the Area of Habitat data.
These range maps were produced by cleaning those supplied as an argument
to \code{x} so they can be used to generate the Area of Habitat data.
Specifically, the object contains the following columns:
\describe{
\item{id_no}{\code{numeric} species' taxon identifier on the IUCN Red List.}
\item{binomial}{\code{character} species name.}
\item{seasonal}{\code{numeric} seasonal distribution code.}
\item{full_habitat_code}{\code{character} all habitat classification
codes that contain suitable habitat for the species.
If a given species has multiple suitable habitat classes,
then these are denoted using a pipe-delimited format.
For example, if the habitat classes denoted with the codes
\code{"1.5"} and \code{"1.9"} were considered suitable for a given species, then
these codes would be indicated as \code{"1.5|1.9"}.}
\item{habitat_code}{\code{character} habitat
codes used to create the species' Area of Habitat data.
Since the argument to \code{habitat_data} may not contain habitat
classes for all suitable habitats for a given species
(e.g. the default dataset does not contain subterranean cave systems),
this column contains the subset of the habitat codes listed in the
\code{"full_habitat_code"} column that were used for processing.}
\item{elevation_lower}{\code{numeric} lower elevation threshold used to create
the species' Area of Habitat data.}
\item{elevation_upper}{\code{numeric} upper elevation threshold used to create
the species' Area of Habitat data.}
\item{xmin}{\code{numeric} value describing the spatial extent of
the Area of Habitat data.}
\item{xmax}{\code{numeric} value describing the spatial extent of
the Area of Habitat data.}
\item{ymin}{\code{numeric} value describing the spatial extent of
the Area of Habitat data.}
\item{ymax}{\code{numeric} value describing the spatial extent of
the Area of Habitat data.}
\item{path}{\code{character} file paths for Area of Habitat data
(in GeoTIFF format).
The Area of Habitat datasets denote the proportion of land
inside different grid cells that contain suitable habitat for
species' seasonal distribution.
Thus a grid cell with a value of 0 has 0\% of its spatial extent covered by
suitable habitat, a value of 1 has 100\% of its spatial extent covered by
suitable habitat, and a missing (\code{NA}) value
means that the cell is located outside of the species'
distribution (per the geographic range data).
File paths that are denoted with missing (\code{NA}) values correspond to
species distributions that were not processed (e.g. due to lack of habitat
data).}
}
}
\description{
Create Area of Habitat (AOH) data for species based on their altitudinal and
habitat preferences (Brooks \emph{et al.} 2019).
Please note that these procedures are designed for terrestrial species
and will not apply to marine or freshwater species.
}
\section{Species range data format}{

Species range data are expected to follow the data format conventions
for the IUCN Red List (see \href{https://www.iucnredlist.org/resources/mappingstandards}{IUCN Red List documentation} for
details). Specifically, the argument to \code{x} must an
\code{\link[sf:sf]{sf::st_sf()}} object with the following columns: \code{id_no} (or \code{SISID}),
\code{presence},
\code{origin}, \code{seasonal}, \code{terrestrial} (or \code{terrestial}), \code{freshwater} and
\code{marine}. Below we provide a brief description of each column:

\describe{

\item{\code{id_no} (or \code{SISID})}{\code{numeric} taxon identifier on the IUCN Red List.}

\item{\code{presence}}{\code{numeric} identifier describing information about
the presence of the taxon in the range data.}

\item{\code{origin}}{\code{numeric}  identifier describing if the species is native
to the location(s) described by the range data.}

\item{\code{seasonality}}{\code{numeric} identifier describing if the species
is occupied by the location(s) describe by the range data throughout
the whole year, of if only during certain seasons.}

\item{\code{terrestial}}{\code{character} value indicating if the range
data pertain to terrestrial environments (with \code{"true"} or \code{"false"}
values.)}

\item{\code{freshwater}}{\code{character} value indicating if the range
data pertain to freshwater environments (with \code{"true"} or \code{"false"}
values.)}

\item{\code{marine}}{\code{character} value indicating if the range
data pertain to marine environments (with \code{"true"} or \code{"false"}
values.)}

}
}

\section{Data processing}{

The Area of Habitat data are produced using the following procedures.

\enumerate{

\item Global elevation and habitat classification are imported if needed.
(see \code{\link[=get_elevation_data]{get_elevation_data()}} and \code{\link[=get_habitat_data]{get_habitat_data()}} for details).
If these data are not available in the cache directory
(i.e. argument to \code{cache_dir}), then they are automatically downloaded
to the cache directory.
For convenience, preprocessed versions of these datasets are downloaded
to reduce processing time.
Note that if elevation and habitat data are manually supplied
(i.e. as arguments to \code{elevation_data} and \code{habitat_data}), then
these datasets are used to generate Area of Habitat data.

\item Species range data cleaned to prepare them for subsequent
analysis. Briefly, this process involves excluding places where the
(i) species' presence is not \emph{extant} or
\emph{probably extant}
(i.e. filtering based on \code{presence == 1} or \code{presence == 2});
(ii) species' origin is not \emph{native},
\emph{reintroduced}, or the result of \emph{assisted colonization}
(i.e. filtering based on \code{origin == 1}, \code{origin == 2}, or \code{origin == 6});
(iii) available information on which species' seasonal distribution
is \emph{uncertain}
(i.e. filtering based on \code{seasonal != 5}); and
(iv) species' distribution is not terrestrial
(i.e. filtering based on where \code{terrestrial == "true"}).
Additionally, the species' range data spatially dissolved so that each
seasonal distribution for each taxon is represented by a separate geometry.
Finally, geoprocessing routines are used to detect and repair
any invalid geometries.

\item Species' altitudinal limit and habitat affiliation data are
imported if needed
(see \code{\link[=get_spp_summary_data]{get_spp_summary_data()}} and \code{\link[=get_spp_habitat_data]{get_spp_habitat_data()}} for details).
If these data are not available in the cache directory
(i.e. argument to \code{cache_dir}), then they are automatically downloaded
to the cache directory
Note that if species' altitudinal limit and habitat affiliation
data are manually supplied (i.e. as arguments to \code{spp_summary_data} and
\code{spp_habitat_data}), then these datasets are used to generate
Area of Habitat data.

\item The species data are collated into a single dataset containing
their geographic ranges, altitudinal limits, and habitat affiliations.
Specifically, taxon identifiers (per the \code{id_no}/\code{SISID} columns)
are used merge the datasets together.
If a species lacks lower or upper altitudinal limits,
then limits of 0 and 9,000 m are assumed respectively
(following Lumbierres \emph{et al.} 2021).
Additionally, the following rules are used to assign
habitat affiliations for species' distributions.
First, habitat affiliations are only
included in the collated dataset if they are classified as \emph{suitable}
or \emph{major}.
Second, if a habitat affiliation is defined for a specific seasonal
distribution of a particular species (e.g. \emph{non-breeding}), then that
habitat affiliation is only assigned to that specific seasonal
distribution for the species.
Third, if a habitat affiliation is not defined for a specific
seasonal distribution, then the habitat is assigned to all seasonal
distributions associated with the species.
Fourth, because the \emph{resident} distributions of some bird species
lack specific habitat affiliation data, the habitat affiliation data
for these \emph{resident} distributions are assigned based on habitat
affiliations for the species' \emph{breeding} distribution.

\item Preliminary geoprocessing routines are used to prepare the habitat and
elevation data for subsequent processing. These routines involve
(i) cropping the template data based on the maximum
extent of the species' ranges;
(ii) aligning the habitat data to conform to the same spatial
properties as the template data; and
(iii) aligning the elevation data to conform to the same spatial
properties as the template data.
Specifically, data alignment procedures involve reprojecting the
the data to same coordinate reference system as the template data
(if needed), and cropping their spatial extent to match
that of the template data (if needed).

\item The Area of Habitat data are then generated for each seasonal
distribution of each species. For a given species' distribution,
the data are generated by
(i) selecting layers from the habitat data that correspond to
the suitable habitats for the species' distribution;
(ii) cropping these layers to the extent of the species'
seasonal distribution;
(iii) summing the layers together to produce a new layer;
(iv) creating a mask based on the altitudinal limits of the species
and the elevation data, and then using the mask to set values
in the new layer to zero if they are outside of the species' altitudinal
limits;
(v) creating a mask by rasterizing the species' seasonal
distribution, and then using the mask to set values in the new
layer to missing (\code{NA}) values if they are outside the species'
distribution;
(vi) clamping the new layer such that values range between
0 and 1,000 (because percent coverage values in the habitat data range
between 0 and 1,000);
(vii) rescaling the new layer to range between 0 and 1; and
(viii) saving the resulting as the Area of Habitat data
for the species' distribution.
Note that species' distributions that already have Area of Habitat data
available in the output directory are skipped.

\item Post-processing routines are used to prepare the returned object.
These routines involve updating the collated species data to include
file names and spatial metadata for the Area of Habitat data.

}
}

\examples{
\dontrun{
# find file path for example range data following IUCN Red List data format
## N.B. the range data were not obtained from the IUCN Red List,
## and were instead based on data from GBIF (https://www.gbif.org/)
path <- system.file("extdata", "EXAMPLE_SPECIES.zip", package = "aoh")

# import data
spp_range_data <- read_spp_range_data(path)

# specify settings for data processing
output_dir <- tempdir()                       # folder to save AOH data
cache_dir <- rappdirs::user_data_dir("aoh")   # persistent storage location
n_threads <- parallel::detectCores() - 1      # speed up analysis

# create cache directory if needed
if (!file.exists(cache_dir)) {
  dir.create(cache_dir, showWarnings = FALSE, recursive = TRUE)
}

# create Area of Habitat data for species
spp_aoh_data <- create_spp_aoh_data(
  x = spp_range_data,
  output_dir = output_dir,
  n_threads = n_threads,
  cache_dir = cache_dir
)
}

\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
# preview data
print(result)
}
\dontshow{\}) # examplesIf}
\dontrun{
# import AOH data as a list of terra::rast() objects
spp_aoh_rasters <- lapply(spp_aoh_data$path, terra::rast)

# print AOH data list
print(spp_aoh_rasters)

# plot the data to visualize the range maps and AOH data
plot_spp_aoh_data(spp_aoh_data)
}
}
\references{
Amatulli G, Domisch S, Tuanmu M-N, Parmentier B, Ranipeta A, Malczyk J, and
Jetz W (2018) A suite of global, cross-scale topographic variables for
environmental and biodiversity modeling. \emph{Scientific Data}, 5:180040.
Available at \url{https://doi.org/10.1038/sdata.2018.40}.

Brooks TM, Pimm SL, Akçakaya HR, Buchanan GM, Butchart SHM, Foden W,
Hilton-Taylor C, Hoffmann M, Jenkins CN, Joppa L, Li BV, Menon V,
Ocampo-Peñuela N, Rondinini C (2019) Measuring terrestrial Area of Habitat
(AOH) and its Utility for the IUCN Red List. \emph{Trends in Ecology & Evolution},
34:977--986. Available at \url{https://doi.org/10.1016/j.tree.2019.06.009}.

Jung M, Dahal PR, Butchart SHM, Donald PF, De Lamo X, Lesiv M, Kapos V,
Rondinini C, and Visconti P (2020a) A global map of
terrestrial habitat types. \emph{Scientific Data}, 7:1--8.
Available at \url{https://doi.org/10.1038/s41597-020-00599-8}.

Jung M, Dahal PR, Butchart SHM, Donald PF, De Lamo X, Lesiv M, Kapos V,
Rondinini C, and Visconti P (2020b) A global map of
terrestrial habitat types (insert version) [Data set].
\emph{Zenodo Digital Repository}.
Available at \url{https://doi.org/10.5281/zenodo.4058819}.
}
