% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_gdal.R
\name{gdal_version}
\alias{gdal_version}
\title{GDAL version}
\usage{
gdal_version()
}
\value{
A \code{character} value describing the version of GDAL installed.
If GDAL is not installed, then a missing (\code{NA}) value is returned.
}
\description{
Find the version of GDAL installed.
}
\examples{
# show version of GDAL installed
print(gdal_version())

}
