% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_elevation_data.R
\name{get_global_elevation_data}
\alias{get_global_elevation_data}
\title{Get global elevation data}
\usage{
get_global_elevation_data(
  dir = tempdir(),
  version = "latest",
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dir}{\code{character} folder path for downloading and caching data.
By default, data are downloaded to a temporary directory (i.e. \code{tempdir()}).
\strong{To avoid downloading the same data multiple times, it is strongly
recommended to specify a persistent storage location (see Examples below).}}

\item{version}{\code{character} indicating the specific version of the dataset
that should be downloaded. The version should be indicated using the
Digital Object Identifier of the specific version required (e.g.
\code{"10.5281/zenodo.3673586"}).
Defaults to \code{"latest"} such that the latest
release of the dataset with available habitat data is used.}

\item{force}{\code{logical} should the data be downloaded even if the
the data are already available?
Defaults to \code{FALSE}.}

\item{verbose}{\code{logical} should progress be displayed while downloading
data?
Defaults to \code{TRUE}.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} object containing the elevation data
(100 m resolution). Pixel values indicate elevation (m).
}
\description{
Import elevation data derived from Robinson \emph{et al.} (2014).
If data are not available locally, they are downloaded from
a \href{https://doi.org/10.5281/zenodo.5719982}{Zenodo repository}.
}
\details{
The data were produced by (i) obtaining raw elevation data from
\href{https://www.earthenv.org/DEM}{EarthEnv project}, (ii) collating
the data into a single raster (GeoTIFF) file, and then (iii)
reprojecting the data to the World Behrmannn coordinate reference
systems (ESRI:54017). For further details, see the
\href{https://github.com/jeffreyhanson/dem}{online repository} containing
code used to process the data.
}
\examples{
\dontrun{
# define persistent storage location
download_dir <- rappdirs::user_data_dir("aoh")

# create download directory if needed
if (!file.exists(download_dir)) {
  dir.create(download_dir, showWarnings = FALSE, recursive = TRUE)
}

# download and import elevation data
elev_data <- get_global_elevation_data(download_dir)

# preview data
print(elev_data)

# plot data
plot(elev_data)
}
}
\references{
Robinson N, Regetz J, and Guralnick RP (2014) EarthEnv-DEM90: A nearly-
global, void-free, multi-scale smoothed, 90m digital elevation model from
fused ASTER and SRTM data.
\emph{ISPRS Journal of Photogrammetry and Remote Sensing}, 87:57--67.
Available at \url{https://doi.org/10.1016/j.isprsjprs.2013.11.002}
}
