% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_gdal.R
\name{is_gdal_available}
\alias{is_gdal_available}
\title{Is GDAL available?}
\usage{
is_gdal_available()
}
\value{
A \code{logical} indicating if GDAL is available or not.
}
\description{
Check if GDAL is available for processing data.
}
\details{
The function checks if GDAL is available by (1) verifying that the
\pkg{gdalUtils} package in installed and (2) verifying that
attempting to query the installed version of GDAL does not produce
an error (i.e. \code{system("gdalinfo --version")}).
}
\examples{
# check if GDAL is available?
print(is_gdal_available())

}
