% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_gdal.R
\name{is_gdal_calc_available}
\alias{is_gdal_calc_available}
\title{Is gdal_calc available?}
\usage{
is_gdal_calc_available()
}
\value{
A \code{logical} value indicating if it is available.
}
\description{
Check if \code{gdal_calc.py} is available.
}
\examples{
# see if gdal_calc is available
print(is_gdal_calc_available())

}
