% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_available.R
\name{is_grass_available}
\alias{is_grass_available}
\title{Is GRASS available?}
\usage{
is_grass_available()
}
\value{
A \code{logical} value indicating if it is available.
}
\description{
Check if the Geographic Resources Analysis Support System (GRASS)
is available for processing data.
}
\details{
The function verifies that
(1) the \pkg{rgrass} package in installed,
(2) the \pkg{link2GI} package in installed,
(3) GRASS is installed (i.e., via \link[link2GI:findGRASS]{link2GI::findGRASS}), and
(4) the version of GRASS installed is at least 7.8.7.
If any of these checks fail, then GRASS is not considered available.
}
\examples{
\dontrun{
# check if GRASS is available?
print(is_grass_available())
}
}
