% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spp_range_data.R
\name{read_spp_range_data}
\alias{read_spp_range_data}
\title{Read species range data}
\usage{
read_spp_range_data(path, n = NULL)
}
\arguments{
\item{path}{\code{character} File path to the data (zip archive) file.}

\item{n}{\code{numeric} Number of features in the dataset to import.
Defaults to \code{NULL} such that all available data is imported.}
}
\value{
A \code{\link[sf:sf]{sf::sf()}} object containing the dataset.
}
\description{
Import species geographic range (i.e. extent of occurrence) data obtained
from the
\href{https://www.iucnredlist.org/}{International Union for Conservation of Nature (IUCN) Red List of Threatened Species}.
}
\details{
Data for amphibians, reptiles, and mammals can be obtained directly from
the International Union for Conservation of Nature (IUCN) Red List website
(see https://www.iucnredlist.org/resources/spatial-data-download).
Data for birds can be obtained by requesting data from
\href{http://www.birdlife.org/}{BirdLife International}
(see \url{http://datazone.birdlife.org/species/requestdis}).
}
\examples{
# find file path for example range data following IUCN Red List data format
## N.B. the range data were not obtained from the IUCN Red List,
## and were instead based on data from GBIF (https://www.gbif.org/)
path <- system.file("extdata", "EXAMPLE_SPECIES.zip", package = "aoh")

# import data
spp_range_data <- read_spp_range_data(path)

# preview data
print(spp_range_data)

# plot data
plot(spp_range_data[, "id_no"])
}
