% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_spp_data.R
\name{simulate_spp_data}
\alias{simulate_spp_data}
\title{Simulate species data}
\usage{
simulate_spp_data(
  n,
  boundary_data,
  habitat_data = NULL,
  elevation_data = NULL,
  crosswalk_data = NULL,
  rf_model_scale_min = 1e+05,
  rf_model_scale_max = 2e+05,
  cache_dir = tempdir(),
  habitat_version = "latest",
  force = FALSE,
  omit_habitat_codes = iucn_habitat_codes_marine(),
  verbose = TRUE
)
}
\arguments{
\item{n}{\code{integer} Number of species to simulate.}

\item{boundary_data}{\code{\link[sf:sf]{sf::st_sf()}} Spatial object delineating the
spatial extent and boundary for simulating species ranges.}

\item{habitat_data}{\code{\link[terra:rast]{terra::rast()}} Raster data indicating the
presence of different habitat classes across world
(e.g., Jung \emph{et al.} 2020a,b; Lumbierres 2021; Lumbierres \emph{et al.} 2021).
Each grid cell should contain an \code{integer} value that specifies which
habitat class is present within the cell
(based on the argument to \code{crosswalk_data}).
Defaults to \code{NULL} such that data are automatically obtained (using
\code{\link[=get_lumbierres_habitat_data]{get_lumbierres_habitat_data()}}).}

\item{elevation_data}{\code{\link[terra:rast]{terra::rast()}} Raster data delineating the
worldwide elevation data (e.g., Robinson \emph{et al.} 2014).
Defaults to \code{NULL} such that data
are automatically obtained (using \code{\link[=get_global_elevation_data]{get_global_elevation_data()}}).
If the data are obtained automatically, then a preprocessed version
of the habitat data will be used to reduce processing time.}

\item{crosswalk_data}{\code{\link[=data.frame]{data.frame()}} Table containing data that indicate
which grid cell values in the argument to \code{habitat_data} correspond to
which IUCN habitat classification codes. The argument should contain
a \code{code} column that specifies a set of IUCN habitat classification
codes (see \code{\link[=iucn_habitat_data]{iucn_habitat_data()}}, and a \code{value} column that specifies
different values in the argument to \code{habitat_data}.
Defaults to \code{NULL} such that the crosswalk for the default habitat
data are used (i.e., \code{\link[=crosswalk_lumbierres_data]{crosswalk_lumbierres_data()}}).}

\item{rf_model_scale_min}{\code{numeric} Minimum scaling parameter used
to control the smallest possible size of simulated species ranges.
See the documentation for the \code{scale} parameter in
\code{RandomFields::RMgauss()} for more information.
Defaults to \code{1e+5}.}

\item{rf_model_scale_max}{\code{numeric} Maximum scaling parameter used
to control the largest possible size of simulated species ranges.
See the documentation for the \code{scale} parameter in
\code{RandomFields::RMgauss()} for more information.
Defaults to \code{2e+5}.}

\item{cache_dir}{\code{character} Folder path for downloading and caching data.
By default, a temporary directory is used (i.e., \code{tempdir()}).
\strong{To avoid downloading the same data multiple times, it is strongly
recommended to specify a persistent storage location (see Examples below).}}

\item{habitat_version}{\code{character} Version of the
habitat dataset that should be used. See documentation for the
the \code{version} parameter in the \code{\link[=get_lumbierres_habitat_data]{get_lumbierres_habitat_data()}} function
for further details.
This parameter is only used if habitat data are obtained
automatically (i.e., the argument to \code{habitat_data} is \code{NULL}).
Defaults to \code{"latest"} such that the most recent version of the dataset is
used if data need to be obtained.}

\item{force}{\code{logical} Should the data be downloaded even if the
the data are already available?
Defaults to \code{FALSE}.}

\item{omit_habitat_codes}{\code{character} Habitat classification codes
to omit from resulting Area of Habitat data.
Please see the \href{https://www.iucnredlist.org/resources/habitat-classification-scheme}{IUCN Red List Habitat Classification Scheme}
for the full range of habitat classification codes.
For example,
if the aim is to identify natural places that contain suitable conditions,
then areas classified as anthropogenically modified
(\code{\link[=iucn_habitat_codes_artificial]{iucn_habitat_codes_artificial()}}),
introduced vegetation (\code{\link[=iucn_habitat_codes_introduced]{iucn_habitat_codes_introduced()}},
or unknown habitat (\code{\link[=iucn_habitat_codes_misc]{iucn_habitat_codes_misc()}}) should
be excluded.
Defaults to \code{\link[=iucn_habitat_codes_marine]{iucn_habitat_codes_marine()}}, such that marine
habitats are excluded.}

\item{verbose}{\code{logical} Should progress be displayed while processing data?
Defaults to \code{TRUE}.}
}
\value{
A \code{list} object containing simulated data that are formatted following
conventions used by the
\href{https://www.iucnredlist.org/}{International Union for Conservation of Nature (IUCN) Red List of Threatened Species}.
It contains the following elements:
\describe{
\item{spp_range_data}{A \code{\link[sf:sf]{sf::st_sf()}} object containing the species'
geographic range data.}
\item{spp_summary_data}{A \code{\link[tibble:tibble]{tibble::tibble()}} object containing summary
information about the species (including altitudinal limit information.}
\item{spp_habitat_data}{A \code{\link[tibble:tibble]{tibble::tibble()}} object containing habitat
preferences for the species.}
}
}
\description{
Simulate species data for creating Area of Habitat data
(Brooks \emph{et al.} 2019). Specifically, data are simulated to define
species geographic ranges, summary information, and habitat preferences.
}
\examples{
# please ensure that the RandomFields and smoothr packages are installed
# to run these examples

\dontshow{if (require(RandomFields) && require(smoothr)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
# define persistent storage location
download_dir <- rappdirs::user_data_dir("aoh")

# create download directory if needed
if (!file.exists(download_dir)) {
  dir.create(download_dir, showWarnings = FALSE, recursive = TRUE)
}

# specify file path for boundary data
boundary_path <- system.file("shape/nc.shp", package = "sf")

# import boundary data to simulate species data
boundary_data <- sf::st_union(sf::read_sf(boundary_path))

# set random number generator seeds for consistency
set.seed(500)
RandomFields::RFoptions(seed = 500)

# simulate data for 5 species
x <- simulate_spp_data(
  n = 5, boundary_data = boundary_data, cache_dir = download_dir
)

# preview species range data
print(x$spp_range_data)

# preview species habitat affiliation data
print(x$spp_habitat_data)

# preview species summary data
print(x$spp_summary_data)
}
\dontshow{\}) # examplesIf}
}
\references{
Brooks TM, Pimm SL, Akçakaya HR, Buchanan GM, Butchart SHM, Foden W,
Hilton-Taylor C, Hoffmann M, Jenkins CN, Joppa L, Li BV, Menon V,
Ocampo-Peñuela N, Rondinini C (2019) Measuring terrestrial Area of Habitat
(AOH) and its Utility for the IUCN Red List. \emph{Trends in Ecology & Evolution},
34:977--986. Available at \url{https://doi.org/10.1016/j.tree.2019.06.009}.
}
\seealso{
See \code{\link[=create_spp_aoh_data]{create_spp_aoh_data()}} for creating Area of Habitat maps using
data for real or simulated species.
}
