% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_spp_data.R
\name{simulate_spp_data}
\alias{simulate_spp_data}
\title{Simulate species data}
\usage{
simulate_spp_data(
  n,
  boundary_data,
  habitat_data = NULL,
  elevation_data = NULL,
  rf_model_scale_min = 1e+05,
  rf_model_scale_max = 2e+05,
  cache_dir = tempdir(),
  habitat_version = "latest",
  force = FALSE,
  omit_habitat_codes = default_omit_iucn_habitat_codes(),
  verbose = TRUE
)
}
\arguments{
\item{n}{\code{integer} Number of species to simulate.}

\item{boundary_data}{\code{\link[sf:sf]{sf::st_sf()}} Spatial object delineating the
spatial extent and boundary for simulating species ranges.}

\item{habitat_data}{\code{\link[terra:rast]{terra::rast()}} Multi-layer raster data delineating the
coverage of different habitat classes across world
(e.g. Jung \emph{et al.} 2020a).
Each layer should correspond to a different habitat class,
and the name of each layer should contain a unique code used to identify
which places contain suitable habitat for a given species
(per the \code{"code"} column in the the argument to \code{spp_habitat_data}).
For each layer, pixel values should indicate the percentage of available
an available habitat class.
\strong{Critically, a value of 1000 should indicate 100\%, and a value of 0
should indicate 0\%.}
Defaults to \code{NULL} such that data
are automatically obtained from Jung \emph{et al.} 2020b
(using \code{\link[=get_global_habitat_data]{get_global_habitat_data()}}).}

\item{elevation_data}{\code{\link[terra:rast]{terra::rast()}} Raster data delineating the
worldwide elevation data (e.g. Jung \emph{et al.} 2020a).
Defaults to \code{NULL} such that data
are automatically obtained from Jung \emph{et al.} 2020b
(using \code{\link[=get_global_elevation_data]{get_global_elevation_data()}}).}

\item{rf_model_scale_min}{\code{numeric} Minimum scaling parameter used
to control the smallest possible size of simulated species ranges.
See the documentation for the \code{scale} parameter in
\code{RandomFields::RMgauss()} for more information.
Defaults to \code{1e+5}.}

\item{rf_model_scale_max}{\code{numeric} Maximum scaling parameter used
to control the largest possible size of simulated species ranges.
See the documentation for the \code{scale} parameter in
\code{RandomFields::RMgauss()} for more information.
Defaults to \code{2e+5}.}

\item{cache_dir}{\code{character} folder path for downloading and caching data.
By default, a temporary directory is used (i.e. \code{tempdir()}).
\strong{To avoid downloading the same data multiple times, it is strongly
recommended to specify a persistent storage location (see Examples below).}}

\item{habitat_version}{\code{character} Version of the
habitat dataset that should be used. See documentation for the
the \code{version} parameter in the \code{\link[=get_global_habitat_data]{get_global_habitat_data()}} function
for further details.
Defaults to \code{"latest"} such that the most recent version of the dataset is
used.}

\item{force}{\code{logical} should the data be downloaded even if the
the data are already available?
Defaults to \code{FALSE}.}

\item{omit_habitat_codes}{\code{character} Habitat classification codes
to omit from resulting Area of Habitat data. If the aim is to identify
places that contain natural, suitable habitats, then processing should
exclude (i) anthropogenically modified habitat classifications and (ii)
unknown habitat classification types. Defaults to all artificial,
introduced vegetation, and unknown habitat classifications on
the \href{https://www.iucnredlist.org/resources/habitat-classification-scheme}{IUCN Red List Habitat Classification Scheme}
(see \code{\link[=default_omit_iucn_habitat_codes]{default_omit_iucn_habitat_codes()}},}

\item{verbose}{\code{logical} Should progress be displayed while processing data?
Defaults to \code{TRUE}.}
}
\value{
A \code{list} object containing simulated data that are formatted following
conventions used by the
\href{https://www.iucnredlist.org/}{International Union for Conservation of Nature (IUCN) Red List of Threatened Species}.
It contains the following elements:
\describe{
\item{spp_range_data}{A \code{\link[sf:sf]{sf::st_sf()}} object containing the species'
geographic range data.}
\item{spp_summary_data}{A \code{\link[tibble:tibble]{tibble::tibble()}} object containing summary
information about the species (including altitudinal limit information.}
\item{spp_habitat_data}{A \code{\link[tibble:tibble]{tibble::tibble()}} object containing habitat
preferences for the species.}
}
}
\description{
Simulate species data for creating Area of Habitat data
(Brooks \emph{et al.} 2019). Specifically, data are simulated to define
species geographic ranges, summary information, and habitat preferences.
}
\examples{
# please ensure that the RandomFields and smoothr packages are installed
# to run these examples

\dontshow{if (require(RandomFields) && require(smoothr)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
# define persistent storage location
download_dir <- rappdirs::user_data_dir("aoh")

# create download directory if needed
if (!file.exists(download_dir)) {
  dir.create(download_dir, showWarnings = FALSE, recursive = TRUE)
}

# specify file path for boundary data
boundary_path <- system.file("shape/nc.shp", package = "sf")

# import boundary data to simulate species data
boundary_data <- sf::st_union(sf::read_sf(boundary_path))

# specify
# simulate data for 5 species
x <- simulate_spp_data(
  n = 5, boundary_data = boundary_data, cache_dir = download_dir
)

# preview species range data
print(x$spp_range_data)

# preview species habitat affiliation data
print(x$spp_habitat_data)

# preview species summary data
print(x$spp_summary_data)
}
\dontshow{\}) # examplesIf}
}
\references{
Brooks TM, Pimm SL, Akçakaya HR, Buchanan GM, Butchart SHM, Foden W,
Hilton-Taylor C, Hoffmann M, Jenkins CN, Joppa L, Li BV, Menon V,
Ocampo-Peñuela N, Rondinini C (2019) Measuring terrestrial Area of Habitat
(AOH) and its Utility for the IUCN Red List. Trends in Ecology & Evolution.
34:977--986. \url{doi:10.1016/j.tree.2019.06.009}
}
\seealso{
See \code{\link[=create_spp_aoh_data]{create_spp_aoh_data()}} for creating Area of Habitat maps using
data for real or simulated species.
}
