% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra_gdal_calc.R
\name{terra_gdal_calc}
\alias{terra_gdal_calc}
\title{GDAL calculate}
\usage{
terra_gdal_calc(
  x,
  expr,
  y = NULL,
  n_threads = 1,
  filename = tempfile(fileext = ".tif"),
  datatype = "FLT4S",
  tiled = FALSE,
  bigtiff = FALSE,
  compress = "LZW",
  verbose = TRUE,
  output_raster = TRUE
)
}
\arguments{
\item{x}{\code{\link[terra:rast]{terra::rast()}} Raster object with source data.}

\item{expr}{\code{character} Value containing expression.}

\item{y}{\code{terra::rast()} Optional raster for calculations.}

\item{n_threads}{\code{integer} Number of computational threads to use
for data processing.
To reduce run time, it is strongly recommended to set this
parameter based on available resources (see Examples section below).
Defaults to 1.}

\item{filename}{\code{character} Filename for output raster.
Defaults to \code{tempfile(fileext = ".tif")}.}

\item{datatype}{\code{character} Value indicating the data type for saving data.
Defaults to \code{"FLT4S"}.}

\item{tiled}{\code{logical} Value indicating if GeoTIFF files should be tiled.
Defaults to \code{FALSE}.}

\item{bigtiff}{\code{logical} Value indicating the data should be stored in
BIGTIFF format.
Defaults to \code{FALSE}.}

\item{compress}{\code{character} Value indicating compression format.
Defaults to \code{"LZW"}.}

\item{verbose}{\code{logical} Should information be displayed during processing?
Defaults to \code{TRUE}.}

\item{output_raster}{\code{logical} Should a raster (\code{\link[terra:rast]{terra::rast()}})
object be returned?
If \code{FALSE} then the file path for the resulting file is returned.
Defaults to \code{TRUE}.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} raster object.
}
\description{
This function is a wrapper for the \code{gdal_calc.py} script.
}
\examples{
# please ensure that the gdalUtils package is installed and
# GDAL system binaries are installed to run this example

\dontshow{if (is_gdal_available() && is_gdal_calc_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# create raster with data
x <- rast(
  ncols = 40, nrows = 40, xmin = -110, xmax = -90, ymin = 40, ymax=60,
  crs = "+proj=longlat +datum=WGS84"
)
values(x) <- seq_len(ncell(x))

# run calculation
y <- terra_gdal_calc(x, "(X < 20) * 1")

# preview result
print(y)
\dontshow{\}) # examplesIf}
}
